% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{add_covariance_variable_terms}
\alias{add_covariance_variable_terms}
\title{Add Individual Covariance Variables As Terms to Formula}
\usage{
add_covariance_variable_terms(f, covariance)
}
\arguments{
\item{f}{(\code{formula})\cr A formula to which covariance structure terms should
be added}

\item{covariance}{(\code{cov_struct})\cr A covariance structure object from which
additional variables should be sourced.}
}
\value{
A new formula with included covariance terms.
}
\description{
Add Individual Covariance Variables As Terms to Formula
}
\keyword{internal}
