% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{add_pattern}
\alias{add_pattern}
\title{Assign missing patterns}
\usage{
add_pattern(data_wide, y_name)
}
\arguments{
\item{data_wide}{wide form data}

\item{y_name}{names of the longitudinal response}
}
\value{
a wide-form data frame with the assigned missing patterns
}
\description{
For wide form of the data, assign the missing pattern indicator
}
