% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{ancova_fit}
\alias{ancova_fit}
\title{fit ANCOVA for complete data}
\usage{
ancova_fit(data_complete_trt, last_name, covar_name, fit_model)
}
\arguments{
\item{data_complete_trt}{complete wide form data of the treatment group}

\item{last_name}{name of the response at the last time point}

\item{covar_name}{name of the covariates}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regression), MASS::rlm (robust regression)}
}
\value{
A list of estimates:
\itemize{
\item{beta_complete_trt}{estimated regression coefficients}
\item{cov_complete_trt}{covariance of the estimated regression coefficients}
}
}
\description{
Fit ANCOVA for the complete data
}
