% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-methods.R
\name{Anova.mmrm}
\alias{Anova.mmrm}
\title{Anova for mmrm models}
\usage{
\method{Anova}{mmrm}(mod, type = c("II", "III", "2", "3"), tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{mod}{(\code{mmrm})\cr the fitted MMRM.}

\item{type}{(\code{string}) type of test, "II", "III", '2', or '3'.}

\item{tol}{(\code{numeric}) threshold blow which values are treated as 0.}
}
\description{
Anova for mmrm models
}
