% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{approx_inference}
\alias{approx_inference}
\title{Variance estimation of likelihood-based inference approximated by delta
method}
\usage{
approx_inference(
  beta_last_ctl,
  beta_last_trt,
  cov4_ctl,
  cov4_trt,
  beta_complete_trt = NULL,
  cov4_complete_trt = NULL,
  obs_pi,
  var_xbar,
  adj_mean,
  n2,
  method
)
}
\arguments{
\item{beta_last_ctl}{regression coefficient at last time point in control}

\item{beta_last_trt}{regression coefficient at last time point in treatment}

\item{cov4_ctl}{covariance of the coefficient at last time point in control}

\item{cov4_trt}{covariance of the coefficient at last time point in treatment}

\item{beta_complete_trt}{estimated regression coefficients for complete data
in treatment}

\item{cov4_complete_trt}{covariance of the estimated regression coefficients
for complete data in treatment}

\item{obs_pi}{marginal observed probability in each group}

\item{var_xbar}{covariance of the covariates}

\item{adj_mean}{overall mean of the covariates}

\item{n2}{number of subjects in treatment}

\item{method}{name of the likelihood-based method. Available: PMM, CBMI}
}
\value{
Estimation results:
\itemize{
\item{llh_est}{point estimate}
\item{var_llh}{variance estimate approximated by delta method}
}
}
\description{
Compute variance estimate of the likelihood-based inference using delta
method
}
