% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_data.R
\name{assert_data}
\alias{assert_data}
\alias{h_assert_one_rec_pt_visit}
\alias{h_assert_rsp_var}
\alias{h_assert_visit_var}
\title{Assertions for Datasets}
\usage{
h_assert_one_rec_pt_visit(vars, data)

h_assert_rsp_var(vars, data)

h_assert_visit_var(vars, data)

assert_data(vars, data)
}
\arguments{
\item{vars}{(\code{list})\cr variables to use.}

\item{data}{(\code{data.frame})\cr data to use.}
}
\value{
Nothing, only fails with an error if assertion does not hold.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

We provide assertion functions to check the input dataset.
}
\section{Functions}{
\itemize{
\item \code{h_assert_one_rec_pt_visit}: asserts single record per patient and visit.

\item \code{h_assert_rsp_var}: assert numeric response variable.

\item \code{h_assert_visit_var}: assert factor visit variable.

\item \code{assert_data}: high-level assertion function to check the dataset.
}}

