% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcva_data}
\alias{bcva_data}
\title{Example Data on BCVA}
\format{
A \code{tibble} with 10,000 rows and 7 variables:
\itemize{
\item \code{USUBJID}: subject ID.
\item \code{VISITN}: visit number (numeric).
\item \code{AVISIT}: visit number (factor).
\item \code{ARMCD}: treatment, \code{TRT} or \code{CTL}.
\item \code{RACE}: 3-category race.
\item \code{BCVA_BL}: BCVA at baseline.
\item \code{BCVA_CHG}: Change in BCVA at study visits.
}
}
\source{
This is an artificial dataset.
}
\usage{
bcva_data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\note{
Measurements of BCVA (best corrected visual acuity) is a measure of how
how many letters a person can read off of an eye chart using corrective
lenses or contacts. This a common endpoint in ophthalmology trials.
}
\keyword{datasets}
