% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cached_mmrm_results}
\alias{cached_mmrm_results}
\title{Cache Data for \code{mmrm} Model Comparison}
\format{
A \code{list} with following elements:
\itemize{
\item \code{conv_time_fev}: Convergence time on FEV data.
\item \code{conv_time_bcva}: Convergence time on BCVA data.
\item \code{rel_diff_ests_tbl_fev}: Relative difference in estimates on FEV data.
\item \code{rel_diff_ests_tbl_bcva}: Relative difference in estimates on BCVA data.
\item \code{conv_rate}: Convergence rate on data with different missing levels.
\item \code{df_missingness}: Summary of missingness on simulated data.
}
}
\source{
This is created based on simulations on FEV data and BCVA data.
}
\usage{
cached_mmrm_results
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
