% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{car_add_mmrm}
\alias{car_add_mmrm}
\title{Register \code{mmrm} For Use With \code{car::Anova}}
\usage{
car_add_mmrm(quietly = FALSE)
}
\arguments{
\item{quietly}{logical: should progress and error messages be suppressed?}
}
\value{
A logical value indicating whether registration was successful.
}
\description{
Register \code{mmrm} For Use With \code{car::Anova}
}
\keyword{internal}
