% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd4}
\alias{cd4}
\title{Repeated CD4 counts data from AIDS clinical trial}
\format{
A data frame with 5036 rows and 6 variables
\describe{
\item{id}{Subject ID}
\item{trt}{Treatment}
\item{age}{Age (years)}
\item{gender}{Gender (1=M, 0=F)}
\item{week}{Week}
\item{logcd4}{log(CD4 count + 1)}
}
}
\source{
\url{https://content.sph.harvard.edu/fitzmaur/ala/cd4.txt}
}
\usage{
cd4
}
\description{
A dataset containing repeated measures of CD4 counts and baseline covariates.
}
\keyword{datasets}
