% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{check}
\alias{check}
\alias{format_check_messages}
\alias{emit_check_results}
\title{Accumulate Check Results}
\usage{
check(severity, cond, ..., envir = parent.frame())

format_check_messages(x, severity)

emit_check_results(envir = parent.frame())
}
\arguments{
\item{severity}{(\code{string})\cr
A severity name to provide, expecting \code{"warn"} or \code{"stop"}}

\item{cond}{(\code{bool})\cr
A condition upon which the message should be emitted}

\item{...}{A message to emit}

\item{envir}{(\code{environment})\cr
Where to look for the \code{check()} accumulator}

\item{x}{(\code{character})\cr
A named character vector. A subset of this vector is formatted if its name
matches the provided severity.}
}
\value{
The condition used to decide whether to emit a message

A formatted \code{character} vector

\code{TRUE} (invisibly) if success, otherwise errors are emitted
}
\description{
Accumulate Check Results
}
\section{Functions}{
\itemize{
\item \code{format_check_messages()}: Format messages accumulated by \code{check()} for printing as warnings or errors.

\item \code{emit_check_results()}: Emit check messages accumulated by calls to \code{check()}

}}
\keyword{internal}
