% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_package_version}
\alias{check_package_version}
\title{Check suggested dependency against version requirements}
\usage{
check_package_version(pkg, ver = c(NA, NA))
}
\arguments{
\item{pkg}{(\code{character(1L)})\cr A package name.}

\item{ver}{(any object of length 2L whose elements can be provided to
\code{numeric_version()})\cr A minimum and maximum (inclusive) version requirement for
interoperability}
}
\value{
A logical (invisibly) indicating whether the loaded package meets
the version requirements. A warning is emitted otherwise.
}
\description{
Check suggested dependency against version requirements
}
\keyword{internal}
