% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{component}
\alias{component}
\title{Component getter for \code{mmrm_tmb} objects}
\usage{
component(
  object,
  name = c("cov_type", "n_theta", "n_subjects", "n_timepoints", "n_obs", "vcov",
    "varcor", "formula", "dataset", "reml", "convergence", "evaluations", "conv_message",
    "call", "theta_est", "beta_est", "x_matrix", "y_vector", "neg_log_lik", "jac_list",
    "theta_vcov"),
  ...
)
}
\arguments{
\item{object}{a \code{mmrm_tmb} object}

\item{name}{of the component to be retrieved}

\item{\dots}{ignored, for method compatibility}
}
\description{
Component getter for \code{mmrm_tmb} objects
}
\details{
Available \code{component()} names are as follows:
\itemize{
\item 'call': low-level function call which generated the model.
\item 'formula': model formula.
\item 'dataset': data set name.
\item 'cov_type': covariance structure type.
\item 'n_theta': number of parameters.
\item 'n_subjects': number of subjects.
\item 'n_timepoints': number of modeled time points.
\item 'n_obs': total number of observations.
\item 'reml': was REML used (ML was used if \code{FALSE}).
\item 'neg_log_lik': negative log likelihood.
\item 'convergence': convergence code from optimizer.
\item 'conv_message': message accompanying the convergence code.
\item 'evaluations': number of function evaluations for optimization.
\item 'vcov': estimated variance-covariance matrix of coefficients.
\item 'varcor': estimated covariance matrix for residuals.
\item 'theta_est': estimated variance parameters.
\item 'beta_est': estimated coefficients.
\item 'theta_vcov':  estimated variance-covariance matrix of variance parameters.
\item 'x_matrix': design matrix used.
\item 'y_vector': response vector used.
\item 'jac_list': Jacobian, see  \code{\link[=h_jac_list]{h_jac_list()}} for details.
}
}
\seealso{
\code{\link[lme4]{getME.merMod}}

\code{\link[glmmTMB]{getME.glmmTMB}}
}
