% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toeplitz.R
\name{corToep}
\alias{corToep}
\title{Toeplitz Correlation Structure}
\usage{
corToep(value = numeric(0), form = ~1, fixed = FALSE)
}
\arguments{
\item{value}{an optional vector of parameter values.
Must have length number of levels - 1.}

\item{form}{a one sided formula of the form ~ t, or ~ t | g,
specifying a time covariate and, optionally,
a grouping factor g.}

\item{fixed}{an optional logical value indicating whether the coefficients
should be allowed to vary in the optimization, or kept fixed
at their initial value. Defaults to FALSE, in which case the
coefficients are allowed to vary.}
}
\description{
This function is a constructor for the \code{corToep} class,
representing a toeplitz correlation structure.
Objects created using this constructor must later be initialized
using the appropriate \code{Initialize} method.
}
