% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{cov_beta_last}
\alias{cov_beta_last}
\title{Coefficient matrix used to get variance of the transformed coefficients}
\usage{
cov_beta_last(beta, p, k)
}
\arguments{
\item{beta}{coefficients from lmer()}

\item{p}{the number of time point}

\item{k}{dimension of the covariates}
}
\value{
Coefficient matrix
}
\description{
Coefficient matrix used to get variance of the transformed coefficients
}
