% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{cov_struct}
\alias{cov_struct}
\title{Define a Covariance Structure}
\usage{
cov_struct(
  type = cov_types(),
  visits,
  subject,
  group = NULL,
  heterogeneous = FALSE
)
}
\arguments{
\item{type}{(\code{string})\cr
The name of the covariance structure type to use. For available options,
see \code{cov_types()}. If a type abbreviation is used that implies
heterogenicity (e.g. \code{cph}) and no value is provided to \code{heterogeneous},
then the heterogenicity is derived from the type name.}

\item{visits}{(\code{character})\cr
A vector of variable names to use for the longitudinal terms of the
covariance structure. Multiple terms are only permitted for the \code{"spatial"}
covariance type.}

\item{subject}{(\code{string})\cr
The name of the variable that encodes a subject identifier.}

\item{group}{(\code{string})\cr
Optionally, the name of the variable that encodes a grouping variable for
subjects.}
}
\value{
A \code{cov_struct} object
}
\description{
Define a Covariance Structure
}
\examples{
cov_struct("csh", "AVISITN", "USUBJID")
cov_struct("spatial", c("VISITA", "VISITB"), group = "GRP", subject = "SBJ")

}
\seealso{
Other cov_struct: 
\code{\link{as.cov_struct}()},
\code{\link{cov_types}()}
}
\concept{cov_struct}
