% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{cov_type_abbr}
\alias{cov_type_abbr}
\title{Retrieve Associated Abbreviated Covariance Structure Type Name}
\usage{
cov_type_abbr(type)
}
\arguments{
\item{type}{(\code{string})\cr either a full name or abbreviate covariance
structure type name to collapse into an abbreviated type.}
}
\value{
The corresponding abbreviated covariance type name.
}
\description{
Retrieve Associated Abbreviated Covariance Structure Type Name
}
\keyword{internal}
