% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{cov_type_name}
\alias{cov_type_name}
\title{Retrieve Associated Full Covariance Structure Type Name}
\usage{
cov_type_name(type)
}
\arguments{
\item{type}{(\code{string})\cr either a full name or abbreviate covariance
structure type name to convert to a long-form type.}
}
\value{
The corresponding abbreviated covariance type name.
}
\description{
Retrieve Associated Full Covariance Structure Type Name
}
\keyword{internal}
