% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{cov_types}
\alias{cov_types}
\title{Retrieve Available Covariance Structures}
\usage{
cov_types()
}
\value{
A character vector of accepted covariance structure type names and
abbreviations
}
\description{
Retrieve Available Covariance Structures
}
\seealso{
Other cov_struct: 
\code{\link{as.cov_struct}()},
\code{\link{cov_struct}()}
}
\concept{cov_struct}
