% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\docType{data}
\name{COV_TYPES_DB}
\alias{COV_TYPES_DB}
\title{Covariance Type Database}
\format{
An object of class \code{data.frame} with 6 rows and 5 columns.
}
\usage{
COV_TYPES_DB
}
\description{
An internal constant for covariance type information.
}
\details{
Contains columns:
\itemize{
\item \code{name}:          The long-form name of the covariance structure type
\item \code{abbr}:          The abbreviated name of the covariance structure type
\item \code{habbr}:         The abbreviated name of the heterogeneous version of a
covariance structure type.
\item \code{heterogeneous}: A logical value indicating whether the covariance
structure has a heterogeneous counterpart.
\item \code{spatial}:       A logical value indicating whether the covariance
structure is spatial.
}
}
\keyword{internal}
