% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{cov_type}
\alias{cov_type}
\alias{cov_type_spatial}
\title{covariance type}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 1.
}
\usage{
cov_type

cov_type_spatial
}
\description{
covariance type
}
\keyword{datasets}
