% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covariance_types}
\alias{covariance_types}
\title{covariance type}
\format{
vector of supported covariance structures. \code{cov_type} for common time points covariance structures,
\code{cov_type_spatial} for spatial covariance structures.
}
\description{
covariance type
}
\details{
abbreviation for covariance structures
\subsection{Common Covariance Structures}{
\itemize{
\item us: "unstructured"
\item toep: "Toeplitz"
\item toeph: "heterogeneous Toeplitz"
\item ar1: "auto-regressive order one"
\item ar1h: "heterogeneous auto-regressive order one"
\item ad: "ante-dependence"
\item adh: "heterogeneous ante-dependence"
\item cs: "compound symmetry"
\item csh: "heterogeneous compound symmetry"
}
}

\subsection{Spatial Covariance structures}{
\itemize{
\item sp_exp: "spatial exponential"
}
}
}
