% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_start}
\alias{default_start}
\title{Default Starting Value}
\usage{
default_start(cov_type, ...)
}
\arguments{
\item{cov_type}{(\code{string})\cr name of the covariance structure.}

\item{...}{additional arguments passed to inner functions.}
}
\value{
A numeric vector of starting values.
}
\description{
Obtain default starting value based on covariance structure.
}
