% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{df_1d_kr}
\alias{df_1d_kr}
\title{Calculation of Kenward-Roger Degrees of Freedom for One-Dimensional Contrast}
\usage{
df_1d_kr(object, contrast, linear = FALSE)
}
\arguments{
\item{object}{\code{mmrm} object}

\item{contrast}{contrast vector}

\item{linear}{\code{logical}\cr whether to use linear Kenward-Roger approximation}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val}.
}
\description{
Calculation of Kenward-Roger Degrees of Freedom for One-Dimensional Contrast
}
\keyword{internal}
