% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{df_md_kr}
\alias{df_md_kr}
\title{obtain Kenward-Roger degree of freedom}
\usage{
df_md_kr(fit, contrast, order = 2L)
}
\arguments{
\item{fit}{\code{mmrm} fit}

\item{contrast}{contrast vector (ncol = 1)}

\item{order}{Kenward-Roger approximation order}
}
\description{
obtain Kenward-Roger degree of freedom
}
\keyword{internal}
