% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{df_md_sat}
\alias{df_md_sat}
\title{Calculation of Satterthwaite Degrees of Freedom for Multi-Dimensional Contrast}
\usage{
df_md_sat(object, contrast)
}
\arguments{
\item{object}{(\code{mmrm})\cr the MMRM fit.}

\item{contrast}{(\code{matrix})\cr numeric contrast matrix, if given a \code{numeric}
then this is coerced to a row vector. Note that this should not include
elements for singular coefficient estimates, i.e. only refer to the
actually estimated coefficients.}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val}.
}
\description{
Calculation of Satterthwaite Degrees of Freedom for Multi-Dimensional Contrast
}
\keyword{internal}
