% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{drop_covariance_terms}
\alias{drop_covariance_terms}
\title{Drop Formula Terms used for Covariance Structure Definition}
\usage{
drop_covariance_terms(f)
}
\arguments{
\item{f}{(\code{formula})\cr a formula from which covariance terms should be
dropped.}
}
\value{
The formula without accepted covariance terms.
}
\description{
Drop Formula Terms used for Covariance Structure Definition
}
\keyword{internal}
