% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{emmeans_support}
\alias{emmeans_support}
\title{Support for \code{emmeans}}
\description{
This package includes methods that allow \code{mmrm} objects to be used
with the \code{emmeans} package. \code{emmeans} computed estimated marginal means
(also called least-square means) for the coefficients of the MMRM.
}
