% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emp_start}
\alias{emp_start}
\title{Empirical Starting Value}
\usage{
emp_start(
  y_vector,
  x_matrix,
  full_frame,
  visit_var,
  subject_var,
  subject_groups,
  cov_type,
  ...
)
}
\arguments{
\item{y_vector}{(\code{numeric}) response variable.}

\item{x_matrix}{(\code{matrix}) design matrix.}

\item{full_frame}{(\code{data.frame}) full data frame used for
model fitting.}

\item{visit_var}{(\code{string})\cr visit variable.}

\item{subject_var}{(\code{string})\cr subject id variable.}

\item{subject_groups}{(\code{factor})\cr subject group assignment.}

\item{cov_type}{(\code{string})\cr covariance structure type.}

\item{...}{not used.}
}
\value{
A numeric vector of starting values.
}
\description{
Obtain empirical start value
for the covariance parameters.
}
\details{
\code{emp_start} supports all non-spatial covariance types.
It uses linear regression to first obtain the coefficients
and uses the residuals to obtain the empirical
variance-covariance matrix. This is then decomposed into
standard deviations and correlations which are mapped to the
appropriate theta parameterization for the specified
covariance type.
}
