% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{expr_cov_struct_calls}
\alias{expr_cov_struct_calls}
\title{Discover Covariance Structure Calls in Expression}
\usage{
expr_cov_struct_calls(expr)
}
\arguments{
\item{expr}{(\code{expression} or \code{quote} or \code{call})\cr
An expression or language object to traverse in search of formula-style
covariance structure specifications.}
}
\value{
A list of outer-most calls within the expression.
}
\description{
Provided an expression, find any calls as candidates for a covariance
structure specification.
}
\examples{
expr_cov_struct_calls(quote(a + b + c(d | f) + g + h(i)))
expr_cov_struct_calls(quote(c(d | f)))

}
\keyword{internal}
