% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fev_data}
\alias{fev_data}
\title{Example Data on FEV1}
\format{
A \code{tibble} with 800 rows and 7 variables:
\itemize{
\item \code{USUBJID}: subject ID.
\item \code{AVISIT}: visit number.
\item \code{ARMCD}: treatment, \code{TRT} or \code{PBO}.
\item \code{RACE}: 3-category race.
\item \code{SEX}: sex.
\item \code{FEV1_BL}: FEV1 at baseline (\%).
\item \code{FEV1}: FEV1 at study visits.
\item \code{WEIGHT}: weighting variable.
}
}
\source{
This is an artificial dataset.
}
\usage{
fev_data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\note{
Measurements of FEV1 (forced expired volume in one second) is a measure
of how quickly the lungs can be emptied. Low levels of FEV1 may indicate
chronic obstructive pulmonary disease (COPD).
}
\keyword{datasets}
