% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_names}
\alias{fill_names}
\title{Complete \code{character} Vector Names From Values}
\usage{
fill_names(x)
}
\arguments{
\item{x}{(\code{character} or \code{list})\cr value whose names should be completed
from element values.}
}
\value{
A named vector or list.
}
\description{
Complete \code{character} Vector Names From Values
}
\keyword{internal}
