% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit the MMRM with Multiple Optimizers if Necessary}
\usage{
fit_model(formula, data, reml = TRUE, optimizer = "automatic", n_cores = 1L)
}
\arguments{
\item{formula}{(\code{formula})\cr the model formula, see \code{\link[=mmrm_tmb]{mmrm_tmb()}} for details.}

\item{data}{(\code{data})\cr the data to be used for the model.}

\item{reml}{(\code{flag})\cr whether restricted maximum likelihood (REML) estimation is used,
otherwise maximum likelihood (ML) is used.}

\item{optimizer}{(\code{string})\cr optimizer to be used to generate the model.}

\item{n_cores}{(\code{count})\cr number of cores which could in principle be used for
parallel computations on Linux or Mac machines.}
}
\value{
The \code{mmrm_fit} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
When setting \code{optimizer = "automatic"}, first the default optimizer
(\code{L-BFGS-B}) is used to fit the model. If that converges, this is returned.
If not, the other available optimizers from \code{\link[=refit_multiple_optimizers]{refit_multiple_optimizers()}} are
tried. If none of the optimizers converge, then the function fails. Otherwise
the best fit is returned.
}
\examples{
fit <- fit_model(
  formula = FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID),
  data = fev_data
)
}
