% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_single_optimizer}
\alias{fit_single_optimizer}
\title{Fitting \code{glmmTMB} Model}
\usage{
fit_single_optimizer(
  formula,
  data,
  start = NULL,
  optimizer = c("L-BFGS-B", "BFGS", "CG", "nlminb")
)
}
\arguments{
\item{formula}{(\code{formula})\cr the \code{glmmTMB} formula.}

\item{data}{(\code{data})\cr the data to be used for the model.}

\item{start}{(\code{list})\cr a list of starting values.}

\item{optimizer}{(\code{character})\cr optimizer to be used to generate the model.}
}
\value{
the fitted glmmTMB object, with additional attributes containing warnings,
messages, errors, optimizer used and convergence status.
}
\description{
This function helps to fit a \code{glmmTMB} model with a single optimizer, while capturing messages and warnings.
}
\examples{
\dontrun{
mod_fit <- fit_single_optimizer(
formula = h_build_formula(vs),
data = dat,
optimizer = "BFGS"
)
attr(mod_fit, "converged")
}
}
