% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{flatten_expr}
\alias{flatten_expr}
\title{Flatten Nested Binary Expressions into a List of Expressions and Operators}
\usage{
flatten_expr(expr)
}
\arguments{
\item{expr}{(\code{language})\cr
A language object to flatten.}
}
\value{
A list of atomic values, symbols, infix operator names and
subexpressions.
}
\description{
Used primarily to support the parsing of covariance structure definitions
from formulas, this function flattens the syntax tree into a hierarchy-less
grammar, allowing for parsing that doesn't abide by R's native operator
precedence.
}
\examples{
flatten_expr(quote(1 + 2 + 3 | 4))

}
\keyword{internal}
