% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{format.cov_struct}
\alias{format.cov_struct}
\title{Format Covariance Structure Object}
\usage{
\method{format}{cov_struct}(x, ...)
}
\arguments{
\item{x}{(\code{cov_struct})\cr a covariance structure object.}

\item{...}{Additional arguments unused.}
}
\value{
A formatted string for \code{x}.
}
\description{
Format Covariance Structure Object
}
