% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{format_symbols}
\alias{format_symbols}
\title{Format Symbol Objects}
\usage{
format_symbols(x)
}
\arguments{
\item{x}{(\code{character}) A vector of variable names.}
}
\value{
A formatted string of comma-separated variables.
}
\description{
For printing, variable names are converted to symbols and deparsed to use R's
built-in formatting of variables that may contain spaces or quote characters.
}
\keyword{internal}
