% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{formula_rhs}
\alias{formula_rhs}
\title{Extract Right-Hand-Side (rhs) from Formula}
\usage{
formula_rhs(f)
}
\arguments{
\item{f}{(\code{formula})\cr A formula.}
}
\value{
A quote or atomic value derived from the right-hand-side of the
formula.
}
\description{
Extract Right-Hand-Side (rhs) from Formula
}
\examples{
formula_rhs(a ~ b + c)
formula_rhs(~ b + c)

}
\keyword{internal}
