% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chol.R
\name{get_chol}
\alias{get_chol}
\title{Obtain Cholesky for new data and theta.}
\usage{
get_chol(fit, data, weights, theta)
}
\arguments{
\item{fit}{(\code{mmrm_fit})}

\item{data}{(\code{data.frame}) input data frame.}

\item{weights}{(\code{numeric}) vector of weights.}

\item{theta}{(\code{numeric}) new theta estimates.}
}
\value{
\code{list} of Cholesky factors.
}
\description{
Use a fitted model to obtain (weighted) Cholesky factors
with new data, weights and theta.
}
