% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{get_pqr}
\alias{get_pqr}
\title{obtain P, Q, R matix for the mmrm fit}
\usage{
get_pqr(
  x,
  subject_zero_inds,
  visits_zero_inds,
  n_subjects,
  subject_n_visits,
  n_visits,
  cov_type,
  is_spatial,
  theta
)
}
\description{
obtain P, Q, R matix for the mmrm fit
}
