% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{h_add_covariance_terms}
\alias{h_add_covariance_terms}
\title{Add Individual Covariance Variables As Terms to Formula}
\usage{
h_add_covariance_terms(f, covariance)
}
\arguments{
\item{f}{(\code{formula})\cr a formula to which covariance structure terms should
be added.}

\item{covariance}{(\code{cov_struct})\cr a covariance structure object from which
additional variables should be sourced.}
}
\value{
A new formula with included covariance terms.
}
\description{
Add Individual Covariance Variables As Terms to Formula
}
\details{
\code{\link[stats:update]{stats::update()}} is used to append the covariance structure and the environment
attribute will not be changed. This ensures the returned formula and the input formula
have the same environment.
}
\keyword{internal}
