% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_Anova_II_Chisq_mmrm}
\alias{h_Anova_II_Chisq_mmrm}
\alias{h_Anova_III_Chisq_mmrm}
\title{Conduct Type-II or Type-III ANOVA with a Chi-squared Test Statistic.}
\usage{
h_Anova_II_Chisq_mmrm(mod, vcov., singular.ok, ...)

h_Anova_III_Chisq_mmrm(mod, vcov., singular.ok, ...)
}
\arguments{
\item{mod}{(\code{mmrm})\cr the fitted MMRM.}

\item{vcov.}{(\code{matrix} or \code{function})\cr the coefficient-covariance matrix or
a function to compute a covariance matrix. This is only used when
\code{test.statistic = "Chisq"}.}

\item{singular.ok}{(\code{flag})\cr flag indicating whether or not to allow models
with aliased coefficients to undergo \code{Chisq} testing. Only used if
\code{test.statistic = "Chisq"}.}

\item{...}{Passed to \code{\link[car:linearHypothesis]{car::linearHypothesis()}} when \code{test.statistic = "Chisq"}.}
}
\value{
\code{anova} object with one row per variable and columns \code{Chisq} (the
Chi-squared test statistic), \code{Df} (degrees of freedom), and \verb{Pr(>Chisq)}
(p-value).
}
\description{
Conduct Type-II or Type-III ANOVA with a Chi-squared Test Statistic.
}
\keyword{internal}
