% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{h_augment_newdata_pred}
\alias{h_augment_newdata_pred}
\title{Augment Data with Prediction Results}
\usage{
h_augment_newdata_pred(x, newdata, se_fit, interval, ...)
}
\arguments{
\item{x}{(\code{mmrm})\cr fit.}

\item{newdata}{(\code{data.frame})\cr data to predict.}

\item{se_fit}{(\code{flag})\cr whether to return standard error of prediction.}

\item{interval}{(\code{string})\cr type of interval.}

\item{...}{passed to \code{\link[=predict.mmrm_tmb]{predict.mmrm_tmb()}}.}
}
\value{
The \code{newdata} as a \code{tibble} with additional columns \code{.fitted},
\code{.lower}, \code{.upper} (if interval is not \code{none}) and \code{.se.fit} (if \code{se_fit}
requested).
}
\description{
This is used in \code{\link[=augment.mmrm]{augment.mmrm()}}.
}
\keyword{internal}
