% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{h_build_formula}
\alias{h_build_formula}
\title{Building Model Formula}
\usage{
h_build_formula(
  vars,
  cor_struct = c("unstructured", "toeplitz", "auto-regressive", "compound-symmetry",
    "ante-dependence")
)
}
\arguments{
\item{vars}{(\code{list})\cr variables to use in the model.}

\item{cor_struct}{(\code{string})\cr specify the covariance structure to use.}
}
\value{
Formula to use in \code{\link[=h_mmrm_tmb]{h_mmrm_tmb()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
vars <- list(
  response = "AVAL", covariates = c("RACE", "SEX"),
  id = "USUBJID", arm = "ARMCD", visit = "AVISIT"
)
h_build_formula(vars, "compound-symmetry")
h_build_formula(vars)
}
