% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-methods.R
\name{h_coef_table}
\alias{h_coef_table}
\title{Coefficients Table for MMRM Fit}
\usage{
h_coef_table(object)
}
\arguments{
\item{object}{(\code{mmrm})\cr model fit.}
}
\value{
Matrix with one row per coefficient and columns
\code{Estimate}, \verb{Std. Error}, \code{df}, \verb{t value} and \verb{Pr(>|t|)}.
}
\description{
This is used by \code{\link[=summary.mmrm]{summary.mmrm()}} to obtain the coefficients table.
}
\keyword{internal}
