% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_confirm_large_levels}
\alias{h_confirm_large_levels}
\title{Ask for Confirmation on Large Visit Levels}
\usage{
h_confirm_large_levels(x)
}
\arguments{
\item{x}{(\code{numeric})\cr number of visit levels.}
}
\value{
Logical value \code{TRUE}.
}
\description{
Ask the user for confirmation if there are too many visit levels
for non-spatial covariance structure in interactive sessions.
}
\keyword{internal}
