% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_construct_full_frame}
\alias{h_construct_full_frame}
\title{Create model data frame for MMRM fit}
\usage{
h_construct_full_frame(formula_parts, data, drop_visit_levels, ignore_response)
}
\arguments{
\item{formula_parts}{(\code{mmrm_tmb_formula_parts})\cr list with formula parts
from \code{\link[=h_mmrm_tmb_formula_parts]{h_mmrm_tmb_formula_parts()}}.}

\item{data}{(\code{data.frame})\cr as returned bu \code{\link[=h_prepare_data]{h_prepare_data()}}.}

\item{drop_visit_levels}{(\code{flag})\cr whether to drop levels for visit variable, if visit variable is a factor.}

\item{ignore_response}{(\code{flag})\cr ignoring the response variable is required when predicting new data where
the response is unknown}
}
\value{
data.frame
}
\description{
Create model data frame for MMRM fit
}
\keyword{internal}
