% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_construct_model_frame_inputs}
\alias{h_construct_model_frame_inputs}
\title{Construct \code{model.frame()} formula and data inputs from mmrm model fit}
\usage{
h_construct_model_frame_inputs(
  formula,
  data,
  include,
  include_choice = c("subject_var", "visit_var", "group_var", "response_var"),
  full
)
}
\arguments{
\item{formula}{mmrm fit object}

\item{data}{optional data frame that will be
passed to \code{model.frame()} or \code{model.matrix()}}

\item{include}{names of variable to include}

\item{full}{deprecated}
}
\value{
list with two named elements: formula, and data
}
\description{
Function returns a named list with three elements:
\itemize{
\item \code{"formula"}: the formula including the columns requested in the \verb{include=} argument.
\item \code{"formula_full"}: the formula including all columns
\item \code{"data"}: a data frame including all columns where factor and
character columns have been processed with \code{h_factor_ref()}.
}

Input formulas are converted from mmrm-style to a style compatible
with default \code{model.frame()} and \code{model.matrix()} methods.

The full formula is returned so we can construct, for example, the
\code{model.frame()} including all columns as well as the requested subset.
The full set is used to identify rows to include in the reduced model frame.
}
\keyword{internal}
