% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_construct_x_matrix}
\alias{h_construct_x_matrix}
\title{Create X matrix for MMRM fit or prediction}
\usage{
h_construct_x_matrix(
  formula_parts,
  full_frame,
  accept_singular,
  check_singular,
  ignore_response
)
}
\arguments{
\item{formula_parts}{(\code{mmrm_tmb_formula_parts})\cr list with formula parts
from \code{\link[=h_mmrm_tmb_formula_parts]{h_mmrm_tmb_formula_parts()}}.}

\item{full_frame}{(\code{data.frame})\cr as returned bu \code{\link[=h_construct_full_frame]{h_construct_full_frame()}}.}

\item{accept_singular}{(\code{flag})\cr whether below full rank design matrices are reduced
to full rank \code{x_matrix} and remaining coefficients will be missing as per
\code{x_cols_aliased}. Otherwise the function fails for rank deficient design matrices.}

\item{check_singular}{(\code{flag})\cr whether to check for singular model matrix;
set to \code{FALSE} when constructing X matrix for prediction}

\item{ignore_response}{(\code{flag})\cr ignoring the response variable is required
when predicting new data where the response is unknown}
}
\value{
data.frame
}
\description{
Create X matrix for MMRM fit or prediction
}
\keyword{internal}
