% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_contr_sum_type3_contrasts}
\alias{h_contr_sum_type3_contrasts}
\title{Construct Preliminary Contrast Matrices for Type III Tests Assuming Sum Contrasts}
\usage{
h_contr_sum_type3_contrasts(object)
}
\arguments{
\item{mod}{(\code{mmrm})\cr the fitted MMRM.}
}
\value{
A \code{list} of contrast matrices, which are just row-wise term subsets of
an overall identity matrix.
}
\description{
Construct Preliminary Contrast Matrices for Type III Tests Assuming Sum Contrasts
}
\keyword{internal}
