% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_covbeta_fun}
\alias{h_covbeta_fun}
\title{Covariance Matrix for Coefficients Given Variance Parameters}
\usage{
h_covbeta_fun(model)
}
\arguments{
\item{model}{(\code{mmrm_tmb})\cr initial model fit.}
}
\value{
Function with argument \code{theta} that calculates the covariance matrix
for the coefficient vector \code{beta}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
formula <- FEV1 ~ RACE + us(AVISIT | USUBJID)
model <- mmrm_tmb(formula, fev_data)
fun <- h_covbeta_fun(model)
fun(model$theta_est)
model$beta_vcov
}
