% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_dataset_sort_all}
\alias{h_dataset_sort_all}
\title{Sort a Data Frame by All Its Columns in Ascending Order}
\usage{
h_dataset_sort_all(data)
}
\arguments{
\item{data}{(\code{data.frame})\cr a data frame to be sorted.}
}
\value{
A data frame. The same as the input \code{data} with columns in the same
order but with rows sorted by the first column, with ties broken by the
second column, and so forth.
}
\description{
Sort a Data Frame by All Its Columns in Ascending Order
}
\keyword{internal}
