% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_default_value}
\alias{h_default_value}
\title{Default Value on NULL}
\usage{
h_default_value(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Return default value when first argument is NULL.
}
\details{
If \code{x} is NULL, returns \code{y}. Otherwise return \code{x}.
}
\keyword{internal}
