% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between-within.R
\name{h_df_1d_bw}
\alias{h_df_1d_bw}
\title{Calculation of Between-Within Degrees of Freedom for One-Dimensional Contrast}
\usage{
h_df_1d_bw(object, contrast)
}
\arguments{
\item{object}{(\code{mmrm})\cr the MMRM fit.}

\item{contrast}{(\code{numeric})\cr contrast vector. Note that this should not include
elements for singular coefficient estimates, i.e. only refer to the
actually estimated coefficients.}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val}.
}
\description{
Used in \code{\link[=df_1d]{df_1d()}} if method is "Between-Within".
}
\keyword{internal}
