% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_within_classic.R
\name{h_df_1d_bwc}
\alias{h_df_1d_bwc}
\title{Calculation of between-within Degrees of Freedom for One-Dimensional Contrast}
\usage{
h_df_1d_bwc(object, contrast)
}
\arguments{
\item{object}{(\code{mmrm})\cr the MMRM fit.}

\item{contrast}{(\code{numeric})\cr contrast vector. Note that this should not include
elements for singular coefficient estimates, i.e. only refer to the
actually estimated coefficients.}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val}.
}
\description{
Calculates the estimate, standard error, degrees of freedom,
t statistic and p-value for one-dimensional contrast. Used in \code{\link[=df_1d]{df_1d()}} if method is
"Between-within".
}
\keyword{internal}
