% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_df_1d_kr}
\alias{h_df_1d_kr}
\title{Calculation of Kenward-Roger Degrees of Freedom for One-Dimensional Contrast}
\usage{
h_df_1d_kr(object, contrast)
}
\arguments{
\item{object}{(\code{mmrm})\cr the MMRM fit.}

\item{contrast}{(\code{numeric})\cr contrast vector. Note that this should not include
elements for singular coefficient estimates, i.e. only refer to the
actually estimated coefficients.}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val}.
}
\description{
Used in \code{\link[=df_1d]{df_1d()}} if method is "Kenward-Roger", i.e. for both
"Kenward-Roger" and "Kenward-Roger-Linear" vcov.
}
\keyword{internal}
