% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_df_1d_list}
\alias{h_df_1d_list}
\title{Creating Results List for One-Dimensional Contrast}
\usage{
h_df_1d_list(est, var, v_num, v_denom)
}
\arguments{
\item{est}{(\code{number})\cr estimate.}

\item{var}{(\code{number})\cr variance of estimate.}

\item{v_num}{(\code{number})\cr numerator for Satterthwaite d.f.}

\item{v_denom}{(\code{number})\cr denominator for Satterthwaite d.f.}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val} (2-sided p-value).
}
\description{
Creating Results List for One-Dimensional Contrast
}
\keyword{internal}
