% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_df_1d_sat_empirical}
\alias{h_df_1d_sat_empirical}
\title{Helper for Calculation of Satterthwaite with Empirical Covariance Matrix}
\usage{
h_df_1d_sat_empirical(object, contrast_matrix)
}
\arguments{
\item{object}{(\code{mmrm})\cr the MMRM fit.}

\item{contrast_matrix}{(\code{matrix})\cr contrast matrix.}
}
\value{
Adjusted degrees of freedom value.
}
\description{
Used in \code{\link[=h_df_1d_sat]{h_df_1d_sat()}} and \code{\link[=h_df_md_sat]{h_df_md_sat()}} if empirical covariance
matrix is used.
}
\keyword{internal}
