% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between-within.R
\name{h_df_bw_calc}
\alias{h_df_bw_calc}
\title{Calculation of Between-Within Degrees of Freedom}
\usage{
h_df_bw_calc(object)
}
\arguments{
\item{object}{(\code{mmrm})\cr the fitted MMRM.}
}
\value{
List with:
\itemize{
\item \code{coefs_between_within} calculated via \code{\link[=h_within_or_between]{h_within_or_between()}}
\item \code{ddf_between}
\item \code{ddf_within}
}
}
\description{
Used in \code{\link[=h_df_1d_bw]{h_df_1d_bw()}} and \code{\link[=h_df_md_bw]{h_df_md_bw()}}.
}
\keyword{internal}
