% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_df_md_from_1d}
\alias{h_df_md_from_1d}
\title{Creating F-Statistic Results from One-Dimensional Contrast}
\usage{
h_df_md_from_1d(object, contrast)
}
\arguments{
\item{object}{(\code{mmrm})\cr model fit.}

\item{contrast}{(\code{numeric})\cr one-dimensional contrast.}
}
\value{
The one-dimensional degrees of freedom are calculated and then
based on that the p-value is calculated.
}
\description{
Creates multi-dimensional result from one-dimensional contrast from \code{\link[=df_1d]{df_1d()}}.
}
\keyword{internal}
