% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_df_md_list}
\alias{h_df_md_list}
\title{Creating Results List for Multi-Dimensional Contrast}
\usage{
h_df_md_list(f_stat, num_df, denom_df)
}
\arguments{
\item{f_stat}{(\code{number})\cr F-statistic.}

\item{num_df}{(\code{number})\cr numerator degrees of freedom.}

\item{denom_df}{(\code{number})\cr denominator degrees of freedom.}
}
\value{
List with \code{num_df}, \code{denom_df}, \code{f_stat} and \code{p_val} (2-sided p-value).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
h_df_md_list(5, 1, 15)
}
