% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between-within.R
\name{h_df_min_bw}
\alias{h_df_min_bw}
\title{Assign Minimum Degrees of Freedom Given Involved Coefficients}
\usage{
h_df_min_bw(bw_calc, is_coef_involved)
}
\arguments{
\item{bw_calc}{(\code{list})\cr from \code{\link[=h_df_bw_calc]{h_df_bw_calc()}}.}

\item{is_coef_involved}{(\code{logical})\cr whether each coefficient is involved
in the contrast.}
}
\value{
The minimum of the degrees of freedom assigned to each involved
coefficient according to its between-within categorization.
}
\description{
Used in \code{\link[=h_df_1d_bw]{h_df_1d_bw()}} and \code{\link[=h_df_md_bw]{h_df_md_bw()}}.
}
\keyword{internal}
