% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{h_df_to_tibble}
\alias{h_df_to_tibble}
\title{Coerce a Data Frame to a \code{tibble}}
\usage{
h_df_to_tibble(data)
}
\arguments{
\item{data}{(\code{data.frame})\cr what to coerce.}
}
\value{
The \code{data} as a \code{tibble}, potentially with a \code{.rownames} column.
}
\description{
This is used in \code{\link[=h_newdata_add_pred]{h_newdata_add_pred()}}.
}
\details{
This is only a thin wrapper around \code{\link[tibble:as_tibble]{tibble::as_tibble()}}, except
giving a useful error message and it checks for \code{rownames} and adds them
as a new column \code{.rownames} if they are not just a numeric sequence as
per the \code{\link[tibble:rownames]{tibble::has_rownames()}} decision.
}
\keyword{internal}
