% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{h_drop_covariance_terms}
\alias{h_drop_covariance_terms}
\title{Drop Formula Terms used for Covariance Structure Definition}
\usage{
h_drop_covariance_terms(f)
}
\arguments{
\item{f}{(\code{formula})\cr a formula from which covariance terms should be
dropped.}
}
\value{
The formula without accepted covariance terms.
}
\description{
Drop Formula Terms used for Covariance Structure Definition
}
\details{
\code{terms} is used and it will preserve the environment attribute.
This ensures the returned formula and the input formula have the same environment.
}
\keyword{internal}
