% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_drop_level}
\alias{h_drop_level}
\title{Drop Levels of a Factor}
\usage{
h_drop_level(x, exclude, ...)
}
\arguments{
\item{x}{(\code{factor}) input vector to drop levels.}
}
\description{
Drop Levels of a Factor
}
