% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_drop_levels}
\alias{h_drop_levels}
\title{Drop Levels from Dataset}
\usage{
h_drop_levels(data, subject_var, visit_var, except)
}
\arguments{
\item{data}{(\code{data.frame}) data to drop levels.}

\item{subject_var}{(\code{character}) subject variable.}

\item{visit_var}{(\code{character}) visit variable.}

\item{except}{(\code{character}) variables to exclude from dropping.}
}
\description{
Drop Levels from Dataset
}
\keyword{internal}
