% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{h_drop_terms}
\alias{h_drop_terms}
\title{Drop Formula Terms with Character}
\usage{
h_drop_terms(f, drops)
}
\arguments{
\item{f}{(\code{formula})\cr a formula to be updated.}

\item{drops}{(\code{character})\cr representation of elements to be removed.}
}
\value{
A new formula with elements in \code{drops} removed.
}
\description{
Drop Formula Terms with Character
}
\details{
Elements in \code{drops} will be removed from the formula, while the environment
of the formula is unchanged. If \code{drops} is \code{NULL} or \code{character(0)}, the formula is
unchanged.
}
\keyword{internal}
