% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_extra_levels}
\alias{h_extra_levels}
\title{Check if a Factor Should Drop Levels}
\usage{
h_extra_levels(x)
}
\arguments{
\item{x}{(\code{vector}) vector to check.}
}
\description{
Check if a Factor Should Drop Levels
}
