% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{h_extract_covariance_terms}
\alias{h_extract_covariance_terms}
\title{Extract Formula Terms used for Covariance Structure Definition}
\usage{
h_extract_covariance_terms(f)
}
\arguments{
\item{f}{(\code{formula})\cr a formula from which covariance terms should be
extracted.}
}
\value{
A list of covariance structure expressions found in \code{f}.
}
\description{
Extract Formula Terms used for Covariance Structure Definition
}
\keyword{internal}
