% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_factor_ref}
\alias{h_factor_ref}
\title{Convert Character to Factor Following Reference}
\usage{
h_factor_ref(x, ref, var_name = vname(x))
}
\arguments{
\item{x}{(\code{character})\cr or (\code{factor}) input.}

\item{ref}{(\code{factor})\cr reference.}

\item{var_name}{(\code{string})\cr variable name of input \code{x}.}
}
\description{
Convert Character to Factor Following Reference
}
\details{
Use \code{ref} to convert x into factor with the same level.
}
\keyword{internal}
