% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_factor_ref_data}
\alias{h_factor_ref_data}
\title{Convert Character to Factor Following Reference in a Data Frame}
\usage{
h_factor_ref_data(x, ref, vars)
}
\arguments{
\item{x}{(\code{data.frame})\cr  input.}

\item{ref}{(\code{data.frame})\cr reference.}

\item{vars}{(\code{character})\cr variable names to convert.}
}
\description{
Convert Character to Factor Following Reference in a Data Frame
}
\details{
Use \code{ref} to convert \code{x} into data frame whose factors
are of the same levels.
}
\keyword{internal}
