% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_first_contain_categorical}
\alias{h_first_contain_categorical}
\title{Check if the Effect is the First Categorical Effect}
\usage{
h_first_contain_categorical(effect, factors, categorical)
}
\arguments{
\item{effect}{(\code{string}) name of the effect.}

\item{factors}{(\code{matrix}) the factor matrix.}

\item{categorical}{(\code{character}) names of the categorical values.}
}
\description{
Check if the Effect is the First Categorical Effect
}
\keyword{internal}
