% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_fits_common_data}
\alias{h_fits_common_data}
\title{Combine the Datasets from \code{mmrm} Fits}
\usage{
h_fits_common_data(fits)
}
\arguments{
\item{fits}{(\code{list})\cr list of \code{mmrm} fits.}
}
\value{
A data frame combining all the common observations among the
datasets underlying the elements of \code{fits}.
}
\description{
Take the data columns used in each \code{mmrm} fit and \code{\link[=merge]{merge()}} them.
}
\details{
All default arguments for \code{\link[=merge]{merge()}} are used, resulting in a "natural join":
the result will only contain the observations found in all datasets.

\code{\link[=droplevels]{droplevels()}} is applied to the final product to prevent extraneous
warnings.
}
\keyword{internal}
