% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{h_free_cores}
\alias{h_free_cores}
\title{Get an approximate number of free cores.}
\usage{
h_free_cores()
}
\value{
The approximate number of free cores, which is an integer between 1 and one less than
the total cores.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This uses the maximum load average at 1, 5 and 15 minutes on Linux and Mac
machines to approximate the number of busy cores. For Windows, the load percentage is
multiplied with the total number of cores.
We then subtract this from the number of all detected cores. One additional core
is not used for extra safety.
}
