% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_generate_new_name}
\alias{h_generate_new_name}
\title{Generate a Name Not Already In an Environment nor Its Parents}
\usage{
h_generate_new_name(x, env)
}
\arguments{
\item{x}{(\code{string})\cr a candidate name.}

\item{env}{(\code{environment})\cr an \link{environment} whose bindings (and whose
parents' bindings) are checked to ensure that they do not include the
returned value.}
}
\value{
A string that does not match any of the bindings of \code{env} nor its
parents.
}
\description{
Alters the user-supplied string \code{x} using \code{\link[=make.names]{make.names()}} with \code{unique = TRUE}
until it is a syntactically valid name not bound to in \code{env} nor its
\link[=parent.env]{parents}.
}
\keyword{internal}
