% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chol.R
\name{h_get_chol}
\alias{h_get_chol}
\title{Obtain Cholesky for new data and theta.}
\usage{
h_get_chol(
  fit,
  theta = NULL,
  data = NULL,
  weights = NULL,
  complete_case = TRUE
)
}
\arguments{
\item{fit}{(\code{mmrm_fit})\cr produced by \code{\link[=mmrm]{mmrm()}}.}

\item{theta}{(\code{numeric})\cr new theta estimates.}

\item{data}{(\code{data.frame})\cr input data frame.}

\item{weights}{(\code{numeric})\cr vector of weights.}

\item{complete_case}{(\code{flag})\cr whether to use complete input \code{data}.}
}
\value{
\code{list} of Cholesky factors.
}
\description{
Use a fitted model to obtain (weighted) Cholesky factors
with new data, weights and theta.
}
\keyword{internal}
