% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_get_contrast}
\alias{h_get_contrast}
\title{Obtain Contrast for Specified Effect}
\usage{
h_get_contrast(
  object,
  effect,
  type = c("II", "III", "2", "3"),
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{object}{(\code{mmrm})\cr the fitted MMRM.}

\item{effect}{(\code{string}) the name of the effect.}

\item{type}{(\code{string}) type of test, "II", "III", '2', or '3'.}

\item{tol}{(\code{numeric}) threshold blow which values are treated as 0.}
}
\value{
A \code{matrix} of the contrast.
}
\description{
This is support function to obtain contrast matrix for type II/III testing.
}
\keyword{internal}
