% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_get_cov_default}
\alias{h_get_cov_default}
\title{Obtain Default Covariance Method}
\usage{
h_get_cov_default(method = c("Satterthwaite", "Kenward-Roger"))
}
\arguments{
\item{method}{(\code{string})\cr degree of freedom method.}
}
\description{
Obtain the default covariance method depending on
the degree of freedom method used.
}
\details{
The default covariance method is different for different degree of freedom method.
If degree of freedom is "Satterthwaite", "Asymptotic" is returned.
If degree of freedom is "Kenward-Roger", then "Kenward-Roger" is returned.
}
\keyword{internal}
