% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_get_covariate_parts}
\alias{h_get_covariate_parts}
\title{Extraction of Covariate Parts from Character Vector}
\usage{
h_get_covariate_parts(covariates)
}
\arguments{
\item{covariates}{(\code{character})\cr specification in the usual way, see examples.}
}
\value{
Character vector of the covariates involved in \code{covariates} specification.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
h_get_covariate_parts(NULL)
h_get_covariate_parts("VISIT:ARM")
h_get_covariate_parts(c("VISIT:ARM", "COUNTRY"))
}
