% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_get_index}
\alias{h_get_index}
\title{Test if the First Vector is Subset of the Second Vector}
\usage{
h_get_index(x, y)
}
\arguments{
\item{x}{(\code{vector}) the first list.}

\item{y}{(\code{vector}) the second list.}
}
\description{
Test if the First Vector is Subset of the Second Vector
}
