% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_get_method}
\alias{h_get_method}
\title{get the method of degree of freedom and covariance}
\usage{
h_get_method(method)
}
\arguments{
\item{method}{(\code{string})\cr method for degree of freedom and covariance}
}
\value{
the method chosen.
}
\description{
get the method of degree of freedom and covariance
}
