% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_get_minimal_fit_data}
\alias{h_get_minimal_fit_data}
\title{Obtain the Minimal Dataset Needed for an \code{mmrm} Fit}
\usage{
h_get_minimal_fit_data(fit)
}
\arguments{
\item{fits}{(\code{mmrm})\cr a fitted \code{mmrm} model.}
}
\value{
A data frame.
}
\description{
Grab the dataset underlying an \code{mmrm} fit and select only the used columns.
}
\details{
Grabs the response variable along with the predictors named in
\code{fit$formula_parts}.
}
\keyword{internal}
