% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_get_prediction_variance}
\alias{h_get_prediction_variance}
\title{Get Prediction Variance}
\usage{
h_get_prediction_variance(object, n_sim, tmb_data)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM.}

\item{n_sim}{(\code{integer})\cr number of replication of sampling.}

\item{tmb_data}{(\code{mmrm_tmb_data})\cr object.}
}
\description{
Get prediction variance with given fit, tmb_data with sampling method
}
\keyword{internal}
