% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_get_sim_per_subj}
\alias{h_get_sim_per_subj}
\title{Get simulated values by patient}
\usage{
h_get_sim_per_subj(mu, nsub, nsim)
}
\arguments{
\item{mu}{(\code{list})\cr object returned by \code{h_get_prediction()}}

\item{nsub}{(\code{integer})\cr number of subjects}

\item{nsim}{(\code{integer})\cr number of values to simulate}
}
\description{
Get simulated values by patient
}
\keyword{internal}
