% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_get_theta_from_cov}
\alias{h_get_theta_from_cov}
\title{Obtain Theta from Covariance Matrix}
\usage{
h_get_theta_from_cov(covariance)
}
\arguments{
\item{covariance}{(\code{matrix}) of covariance matrix values.}
}
\description{
Obtain unstructured theta from covrariance matrix.
}
\keyword{internal}
