% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_jac_col_as_matrix}
\alias{h_jac_col_as_matrix}
\title{Formatting a Column from Jacobian Matrix as Matrix}
\usage{
h_jac_col_as_matrix(jac_matrix, col)
}
\arguments{
\item{jac_matrix}{(\code{matrix})\cr full Jacobian matrix.}

\item{col}{(\code{int})\cr column index.}
}
\value{
The column \code{col} of \code{jac_matrix} as a square matrix.
Here the values in the column are used to fill the result column
by column.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
jac_matrix <- matrix(1:81, 9, 9)
h_jac_col_as_matrix(jac_matrix, 5)
}
