% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_md_denom_df}
\alias{h_md_denom_df}
\title{Calculating Denominator Degrees of Freedom for the Multi-Dimensional Case}
\usage{
h_md_denom_df(t_stat_df)
}
\arguments{
\item{t_stat_df}{(\code{numeric})\cr \code{n} t-statistic derived degrees of freedom.}
}
\value{
Usually the calculation is returning \code{2 * E / (E - n)} where
\code{E} is the sum of \code{t / (t - 2)} over all \code{t_stat_df} values \code{t}.
}
\description{
Calculates the degree of freedom for multi-dimensional contrast.
}
\note{
If the input values are two similar to each other then just the average
of them is returned. If any of the inputs is not larger than 2 then 2 is
returned.
}
\keyword{internal}
