% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb}
\alias{h_mmrm_tmb}
\title{Fitting an MMRM with \code{TMB}}
\usage{
h_mmrm_tmb(formula, data, reml = TRUE, control = h_mmrm_tmb_control())
}
\arguments{
\item{formula}{(\code{formula})\cr model formula with exactly one special term
specifying the visits within subjects, see details.}

\item{data}{(\code{data.frame})\cr input data containing the variables used
in \code{formula}.}

\item{reml}{(\code{flag})\cr whether restricted maximum likelihood (REML) estimation is used,
otherwise maximum likelihood (ML) is used.}

\item{control}{(\code{mmrm_tmb_control})\cr list of control options produced
by \code{\link[=h_mmrm_tmb_control]{h_mmrm_tmb_control()}}.}
}
\value{
List pf class \code{mmrm_tmb} with:
\itemize{
\item \code{cov}: estimated covariance matrix.
\item \code{beta_est}: vector of coefficient estimates.
\item \code{beta_vcov}: Variance-covariance matrix for coefficient estimates.
\item \code{theta_est}: vector of variance parameter estimates.
\item \code{theta_vcov}: variance-covariance matrix for variance parameter estimates.
\item \code{neg_log_lik}: obtained negative log-likelihood.
\item \code{formula_parts}: input.
\item \code{data}: input.
\item \code{reml}: input as a flag.
\item \code{opt_details}: list with optimization details including convergence code.
\item \code{tmb_object}: original \code{TMB} object created with \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
\item \code{tmb_data}: input.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is the low-level function to fit an MMRM. Note that this does not
try different optimizers or adds Jacobian information etc. in contrast to
\code{\link[=mmrm]{mmrm()}}.
}
\details{
The \code{formula} typically looks like:
\code{FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)}
so specifies response and covariates as usual, and exactly one special term
defines which covariance structure is used and what are the visit and
subject variables.
}
\examples{
formula <- FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
data <- fev_data
system.time(result <- h_mmrm_tmb(formula, data))
}
