% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_assert_fit}
\alias{h_mmrm_tmb_assert_fit}
\title{Asserting and Checking \code{TMB} Fit Result}
\usage{
h_mmrm_tmb_assert_fit(tmb_object, tmb_fit)
}
\arguments{
\item{tmb_object}{(\code{list})\cr created with \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}

\item{tmb_fit}{(\code{list})\cr created with optimizer function.}
}
\value{
Nothing, only used to generate messages, warnings or errors.
}
\description{
Asserting and Checking \code{TMB} Fit Result
}
