% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_assert_opt}
\alias{h_mmrm_tmb_assert_opt}
\title{Asserting and Checking \code{TMB} Optimization Result}
\usage{
h_mmrm_tmb_assert_opt(tmb_object, tmb_opt)
}
\arguments{
\item{tmb_object}{(\code{list})\cr created with \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}

\item{tmb_opt}{(\code{list})\cr optimization result.}
}
\value{
Nothing, only used to generate messages, warnings or errors.
}
\description{
Asserting and Checking \code{TMB} Optimization Result
}
\keyword{internal}
