% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_control}
\alias{h_mmrm_tmb_control}
\title{Control parameters for \code{TMB} fit}
\usage{
h_mmrm_tmb_control(
  optimizer = stats::nlminb,
  optimizer_args = list(),
  optimizer_control = list()
)
}
\arguments{
\item{optimizer}{(\code{function})\cr optimization function.}

\item{optimizer_args}{(\code{list})\cr additional arguments to be passed to optimizer.}

\item{optimizer_control}{(\code{list})\cr specific \code{control} argument for optimizer.}
}
\value{
List with the control parameters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
h_mmrm_tmb_control(
  optimizer = stats::optim,
  optimizer_args = list(method = "L-BFGS-B")
)
}
