% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_extract_terms}
\alias{h_mmrm_tmb_extract_terms}
\title{Extract group/subject vars from covariance term}
\usage{
h_mmrm_tmb_extract_terms(cov_content)
}
\arguments{
\item{cov_content}{(\code{call})\cr covariance term.}
}
\value{
Named list with elements:
\itemize{
\item \code{visit_var}: \code{character} with the visit variable name.
\item \code{subject_var}: \code{string} with the subject variable name.
\item \code{group_var}: \code{string} with the group variable name. If no group specified, this element
is NULL.
}
}
\description{
Extract group/subject vars from covariance term
}
\details{
a covariance term, which takes the following form:
\code{us(time | group / subject)}, or \code{sp_exp(time1, time2, ... | group / subject)}.
Only for spatial covariance structure can have multiple time coordinates.
This function, take the call of \code{time | group / subject}, last element of the covariance term,
extract the corresponding grouping var, time var and subject var.
}
\keyword{internal}
