% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_extract_vars}
\alias{h_mmrm_tmb_extract_vars}
\title{Processing the covariance term}
\usage{
h_mmrm_tmb_extract_vars(cov_term)
}
\arguments{
\item{cov_term}{(\code{call})\cr covariance term.}
}
\value{
Named list of with elements:
\itemize{
\item \code{visit_var}: \code{character} with the visit variable name.
\item \code{subject_var}: \code{string} with the subject variable name.
\item \code{group_var}: \code{string} with the group variable name. If no group specified, this element
is NULL.
\item \code{is_spatial}: \code{logical} indicator of whether the covariance structure is spatial
}
}
\description{
Processing the covariance term
}
\keyword{internal}
