% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_formula_parts}
\alias{h_mmrm_tmb_formula_parts}
\title{Processing the Formula for \code{TMB} Fit}
\usage{
h_mmrm_tmb_formula_parts(formula)
}
\arguments{
\item{formula}{(\code{formula})\cr original formula.}
}
\value{
List of class \code{mmrm_tmb_formula_parts} with elements:
\itemize{
\item \code{model_formula}: \code{formula} with the correlation term is removed.
\item \code{full_formula}: same as \code{model_formula} but includes the \code{subject_var}.
\item \code{corr_type}: \code{string} with correlation term type (e.g. \code{"us"}).
\item \code{visit_var}: \code{string} with the visit variable name.
\item \code{subject_var}: \code{string} with the subject variable name.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
formula <- FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
h_mmrm_tmb_formula_parts(formula)
}
