% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_parameters}
\alias{h_mmrm_tmb_parameters}
\title{Start Parameters for \code{TMB} Fit}
\usage{
h_mmrm_tmb_parameters(formula_parts, tmb_data, start_values)
}
\arguments{
\item{formula_parts}{(\code{mmrm_tmb_formula_parts})\cr produced by
\code{\link[=h_mmrm_tmb_formula_parts]{h_mmrm_tmb_formula_parts()}}.}

\item{tmb_data}{(\code{mmrm_tmb_data})\cr produced by \code{\link[=h_mmrm_tmb_data]{h_mmrm_tmb_data()}}.}

\item{start_values}{(\code{numeric} or \code{NULL})\cr optional start values for variance
parameters.}
}
\value{
List with element \code{theta} containing the start values for the variance
parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
formula <- FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
formula_parts <- h_mmrm_tmb_formula_parts(formula)
tmb_data <- h_mmrm_tmb_data(formula_parts, fev_data, reml = FALSE)
pars <- h_mmrm_tmb_parameters(formula_parts, tmb_data, NULL)
}
