% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_obtain_lvls}
\alias{h_obtain_lvls}
\title{Obtain Levels Prior and Posterior}
\usage{
h_obtain_lvls(var, additional_vars, xlev, factors)
}
\arguments{
\item{var}{(\code{string}) name of the effect.}

\item{additional_vars}{(\code{character}) names of additional variables.}

\item{xlev}{(\code{list}) named list of character levels.}

\item{factors}{(\code{matrix}) the factor matrix.}
}
\description{
Obtain Levels Prior and Posterior
}
\keyword{internal}
