% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_optimizer_fun}
\alias{h_optimizer_fun}
\title{Obtain Optimizer Function with Character}
\usage{
h_optimizer_fun(optimizer = c("L-BFGS-B", "BFGS", "CG", "nlminb"))
}
\arguments{
\item{optimizer}{(\code{character})\cr vector of optimizers.}
}
\value{
A (\code{list})\cr of optimizer functions generated from \code{\link[=h_partial_fun_args]{h_partial_fun_args()}}.
}
\description{
Obtain the optimizer function through the character provided.
}
\keyword{internal}
