% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_prepare_data}
\alias{h_prepare_data}
\title{Preprocess input data for MMRM fit}
\usage{
h_prepare_data(formula_parts, data, weights = NULL)
}
\arguments{
\item{formula_parts}{(\code{mmrm_tmb_formula_parts})\cr list with formula parts
from \code{\link[=h_mmrm_tmb_formula_parts]{h_mmrm_tmb_formula_parts()}}.}

\item{data}{(\code{data.frame})\cr which contains variables used in \code{formula_parts}.}

\item{weights}{(\code{vector})\cr weights to be used in the fitting process.}
}
\value{
data.frame
}
\description{
Preprocess input data for MMRM fit
}
\keyword{internal}
