% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-methods.R
\name{h_print_aic_list}
\alias{h_print_aic_list}
\title{Printing AIC and other Model Fit Criteria}
\usage{
h_print_aic_list(aic_list, digits = 1)
}
\arguments{
\item{aic_list}{(\code{list})\cr list as part of from \code{\link[=summary.mmrm]{summary.mmrm()}}.}

\item{digits}{(\code{number})\cr number of decimal places used with \code{\link[=round]{round()}}.}
}
\description{
This is used in \code{\link[=print.summary.mmrm]{print.summary.mmrm()}}.
}
\keyword{internal}
