% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-methods.R
\name{h_print_call}
\alias{h_print_call}
\title{Printing MMRM Function Call}
\usage{
h_print_call(call, n_obs, n_subjects, n_timepoints)
}
\arguments{
\item{call}{(\code{call})\cr original \code{\link[=mmrm]{mmrm()}} function call.}

\item{n_obs}{(\code{int})\cr number of observations.}

\item{n_subjects}{(\code{int})\cr number of subjects.}

\item{n_timepoints}{(\code{int})\cr number of timepoints.}
}
\description{
This is used in \code{\link[=print.summary.mmrm]{print.summary.mmrm()}}.
}
\keyword{internal}
