% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-methods.R
\name{h_print_cov}
\alias{h_print_cov}
\title{Printing MMRM Correlation Type}
\usage{
h_print_cov(cov_type, n_theta)
}
\arguments{
\item{cov_type}{(\code{string})\cr covariance structure abbreviation.}

\item{n_theta}{(\code{int})\cr number of variance parameters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is used in \code{\link[=print.summary.mmrm]{print.summary.mmrm()}}.
}
