% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_qr}
\alias{h_qr}
\title{Validation with QR Decomposition}
\usage{
h_qr(x, msg)
}
\arguments{
\item{x}{(\code{list})\cr or (\code{matrix})\cr input to be checked.}

\item{msg}{(\code{string})\cr message to raise on error.}
}
\description{
Validation with QR Decomposition
}
\keyword{internal}
