% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_quad_form_vec}
\alias{h_quad_form_vec}
\title{Quadratic Form Calculation}
\usage{
h_quad_form_vec(x, mat)
}
\arguments{
\item{x}{(\code{numeric})\cr interpreted as a row vector.}

\item{mat}{(\code{matrix})\cr square matrix with the same dimensions as \code{x}.}
}
\value{
The number \code{x \%*\% mat \%*\% t(x)} as a numeric (not a matrix).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
h_quad_form_vec(1:2, matrix(1:4, 2, 2))
}
