% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{h_reconcile_cov_struct}
\alias{h_reconcile_cov_struct}
\title{Reconcile Possible Covariance Structure Inputs}
\usage{
h_reconcile_cov_struct(formula = NULL, covariance = NULL)
}
\arguments{
\item{formula}{(\code{formula})\cr the model formula, see details.}

\item{covariance}{(\code{cov_struct})\cr a covariance structure type definition
as produced with \code{\link[=cov_struct]{cov_struct()}}, or value that can be coerced to a
covariance structure using \code{\link[=as.cov_struct]{as.cov_struct()}}. If no value is provided,
a structure is derived from the provided formula.}
}
\value{
The value \code{covariance} if it's provided or a covariance structure
derived from the provided \code{formula} otherwise. An error is raised of both
are provided.
}
\description{
Reconcile Possible Covariance Structure Inputs
}
\keyword{internal}
