% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_register_s3}
\alias{h_register_s3}
\title{Register S3 Method
Register S3 method to a generic.}
\usage{
h_register_s3(pkg, generic, class, envir = parent.frame())
}
\arguments{
\item{pkg}{(\code{string}) name of the package name.}

\item{generic}{(\code{string}) name of the generic.}

\item{class}{(\code{string}) class name the function want to dispatch.}

\item{envir}{(\code{environment}) the location the method is defined.}
}
\description{
Register S3 Method
Register S3 method to a generic.
}
\details{
This function is adapted from \code{emmeans:::register_s3_method()}.
}
\keyword{internal}
