% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_relatives}
\alias{h_relatives}
\title{Get Indices of a Model Term's "Relatives"}
\usage{
h_relatives(term, names, factors)
}
\arguments{
\item{term}{(\code{string})\cr the model term at hand.}

\item{names}{(\code{character})\cr a vector of all the terms in the model.}

\item{factors}{(\code{matrix})\cr the \code{factors} attribute of a \code{\link{terms.object}}.}
}
\description{
For the model term \code{term}, return the indices of all \emph{other} model terms
whose variables contain all of \code{term}'s variables. These are probably most
often the indices of interaction terms whose components contain all \code{term}'s
variables.
}
\keyword{internal}
