% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_residuals_normalized}
\alias{h_residuals_normalized}
\title{Calculate normalized residuals}
\usage{
h_residuals_normalized(object)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM.}
}
\value{
Vector of residuals
}
\description{
This is used by \code{\link[=residuals.mmrm_tmb]{residuals.mmrm_tmb()}} to calculate normalized / scaled residuals.
}
\keyword{internal}
