% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_residuals_pearson}
\alias{h_residuals_pearson}
\title{Calculate Pearson Residuals}
\usage{
h_residuals_pearson(object, resids_unscaled)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM.}

\item{resids_unscaled}{(\code{numeric})\cr the response residuals.}
}
\value{
Vector of residuals.
}
\description{
This is used by \code{\link[=residuals.mmrm_tmb]{residuals.mmrm_tmb()}} to calculate Pearson residuals.
}
\keyword{internal}
