% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_residuals_response}
\alias{h_residuals_response}
\title{Calculate response residuals.}
\usage{
h_residuals_response(object)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM.}
}
\value{
Vector of residuals
}
\description{
This is used by \code{\link[=residuals.mmrm_tmb]{residuals.mmrm_tmb()}} to calculate response residuals.
}
\keyword{internal}
