% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_split_control}
\alias{h_split_control}
\title{Split Control List}
\usage{
h_split_control(control, ...)
}
\arguments{
\item{control}{(\code{mmrm_control})\cr object.}

\item{...}{additional parameters to update the \code{control} object.}
}
\value{
A \code{list} of \code{mmrm_control} entries.
}
\description{
Split the \code{\link[=mmrm_control]{mmrm_control()}} object according to its optimizers and use additional arguments
to replace the elements in the original object.
}
\keyword{internal}
