% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{h_summarize_all_fits}
\alias{h_summarize_all_fits}
\title{Summarizing List of Fits}
\usage{
h_summarize_all_fits(all_fits)
}
\arguments{
\item{all_fits}{(\code{list} of \code{mmrm_fit} or \code{try-error})\cr list of fits.}
}
\value{
List with \code{warnings}, \code{messages}, \code{log_liks} and \code{converged} results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
mod_fit <- fit_single_optimizer(
  formula = FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID),
  data = fev_data,
  optimizer = "nlminb"
)
mod_fit2 <- fit_single_optimizer(
  formula = FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID),
  data = fev_data,
  optimizer = "L-BFGS-B"
)
all_fits <- list(mod_fit, mod_fit2)
h_summarize_all_fits(all_fits)
}
