% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{h_tbl_confint_terms}
\alias{h_tbl_confint_terms}
\title{Extract \code{tibble} with Confidence Intervals and Term Names}
\usage{
h_tbl_confint_terms(x, ...)
}
\arguments{
\item{x}{(\code{mmrm})\cr fit object.}

\item{...}{passed to \code{\link[stats:confint]{stats::confint()}}, hence not used at the moment.}
}
\value{
A \code{tibble} with \code{term}, \code{conf.low}, \code{conf.high} columns.
}
\description{
This is used in \code{\link[=tidy.mmrm]{tidy.mmrm()}}.
}
\keyword{internal}
