% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{h_test_1d}
\alias{h_test_1d}
\title{Creating T-Statistic Test Results For One-Dimensional Contrast}
\usage{
h_test_1d(object, contrast, df)
}
\arguments{
\item{object}{(\code{mmrm})\cr the MMRM fit.}

\item{contrast}{(\code{numeric})\cr contrast vector. Note that this should not include
elements for singular coefficient estimates, i.e. only refer to the
actually estimated coefficients.}

\item{df}{(\code{number})\cr degrees of freedom for the one-dimensional contrast.}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val} (2-sided p-value).
}
\description{
Creates a list of results for one-dimensional contrasts using
a t-test statistic and the given degrees of freedom.
}
\keyword{internal}
