% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_tmb_version_sufficient}
\alias{h_tmb_version_sufficient}
\title{Predicate if the TMB Version Used to Compile the Package is Sufficient}
\usage{
h_tmb_version_sufficient()
}
\value{
Flag whether the TMB version is sufficient.
}
\description{
Predicate if the TMB Version Used to Compile the Package is Sufficient
}
\keyword{internal}
