% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_tmb_warn_optimization}
\alias{h_tmb_warn_optimization}
\title{Warn if TMB is Configured to Optimize Instantly}
\usage{
h_tmb_warn_optimization()
}
\value{
No return value, called for side effects.
}
\description{
This function checks the TMB configuration for the \code{optimize.instantly} setting.
If it is set to \code{TRUE}, a warning is issued indicating that this may lead to
unreproducible results.
}
\keyword{internal}
