% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_tr}
\alias{h_tr}
\title{Trace of a Matrix}
\usage{
h_tr(x)
}
\arguments{
\item{x}{(\code{matrix})\cr square matrix input.}
}
\value{
The trace of the square matrix.
}
\description{
Obtain the trace of a matrix if the matrix is diagonal, otherwise raise an error.
}
\keyword{internal}
