% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_type2_contrast}
\alias{h_type2_contrast}
\title{Obtain Type 2 Contrast for One Specified Effect}
\usage{
h_type2_contrast(object, effect, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{object}{(\code{mmrm})\cr the fitted MMRM.}

\item{effect}{(\code{string}) the name of the effect.}

\item{tol}{(\code{numeric}) threshold blow which values are treated as 0.}
}
\value{
A \code{matrix} of the contrast.
}
\description{
This is support function to obtain contrast matrix for type II testing.
}
\keyword{internal}
