% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{h_type3_contrasts}
\alias{h_type3_contrasts}
\title{Obtain Type 3 Contrast for All Effects}
\usage{
h_type3_contrasts(object, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{object}{(\code{mmrm})\cr the fitted MMRM.}

\item{tol}{(\code{numeric}) threshold below which values are treated as 0.}
}
\value{
A \code{list} of contrast matrices, one per effect.
}
\description{
This is support function to obtain contrast matrices for type III testing.
}
\keyword{internal}
