% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_valid_fit}
\alias{h_valid_fit}
\title{Validate an mmrm Fit}
\usage{
h_valid_fit(object)
}
\arguments{
\item{object}{(\code{mmrm})\cr object.}
}
\description{
Validate an mmrm Fit
}
\details{
During mmrm fitting, sometimes the fit itself converged but
the fit is not valid. This function asserts the fit is valid
through the covariance and coefficient covariance.
}
\keyword{internal}
