% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_valid_formula}
\alias{h_valid_formula}
\title{Validate mmrm Formula}
\usage{
h_valid_formula(formula)
}
\arguments{
\item{formula}{(\code{formula})\cr to check.}
}
\description{
Validate mmrm Formula
}
\details{
In mmrm models, \code{.} is not allowed as it introduces ambiguity of covariates
to be used, so it is not allowed to be in formula.
}
\keyword{internal}
