% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_var_adj_contracted}
\alias{h_var_adj_contracted}
\title{Obtain the Contracted Linear Adjusted Covariance Matrix}
\usage{
h_var_adj_contracted(v, w, p, s_q)
}
\arguments{
\item{v}{(\code{matrix})\cr unadjusted covariance matrix.}

\item{w}{(\code{matrix})\cr covariance matrix of the estimated covariance parameters.}

\item{p}{(\code{matrix})\cr P matrix from \code{\link[=h_get_kr_comp]{h_get_kr_comp()}}.}

\item{s_q}{(\code{matrix})\cr contracted Q sum \code{S_Q} from \code{\link[=h_get_kr_comp]{h_get_kr_comp()}} called with \code{w}.}
}
\value{
The matrix of adjusted covariance matrix.
}
\description{
Obtains the linear Kenward-Roger adjusted covariance matrix for the
coefficient estimates from the \code{P} matrices and the contracted \code{S_Q}.
Used in \code{\link[=mmrm]{mmrm()}} fitting if vcov is "Kenward-Roger-Linear".
}
\details{
See the section "Contracted linear covariance adjustment" in
\code{vignette("kenward", package = "mmrm")}. The full \code{w} is used, so that
cross-group covariance parameter covariances contribute.
}
\keyword{internal}
