% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between-within.R
\name{h_within_or_between}
\alias{h_within_or_between}
\title{Determine Within or Between for each Design Matrix Column}
\usage{
h_within_or_between(x_matrix, subject_ids)
}
\arguments{
\item{x_matrix}{(\code{matrix})\cr the design matrix with column names.}

\item{subject_ids}{(\code{factor})\cr the subject IDs.}
}
\value{
Character vector with "intercept", "within" or "between" for each
design matrix column identified via the names of the vector.
}
\description{
Used in \code{\link[=h_df_bw_calc]{h_df_bw_calc()}} to determine whether a variable
differs only between subjects or also within subjects.
}
\keyword{internal}
