% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{j2r_imp}
\alias{j2r_imp}
\title{J2R imputation}
\usage{
j2r_imp(
  data_wide,
  beta_imp_ctl,
  beta_imp_trt,
  sigma_ctl,
  sigma_trt,
  covar_name,
  y_name,
  trt_name
)
}
\arguments{
\item{data_wide}{wide form of the data}

\item{beta_imp_ctl}{M sets of the estimated MMRM coefficients used in MI
for control group}

\item{beta_imp_trt}{M sets of the estimated MMRM coefficients used in MI
for treatment group}

\item{sigma_ctl}{Estimated population covariance matrix of the longitudinal
response for the control group}

\item{sigma_trt}{Estimated population covariance matrix of the longitudinal
response for the treatment group}

\item{covar_name}{name of the baseline covariates}

\item{y_name}{name of the longitudinal response}

\item{trt_name}{name of the treatment}
}
\value{
imputation set at last time point
}
\description{
Conduct J2R imputation, using the estimated coefficients from MMRM
}
