% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{j2r_llh}
\alias{j2r_llh}
\title{Likelihood-based method with estimation from delta method}
\usage{
j2r_llh(mmrm_formula, data_long, time_name, id_name, trt_name, method = "PMM")
}
\arguments{
\item{mmrm_formula}{formula for fixed effect MMRM under MAR}

\item{data_long}{long-form of the data frame}

\item{time_name}{name of the time variable}

\item{id_name}{name of the id variable}

\item{trt_name}{name of the treatment variable}

\item{method}{specific method for likelihood-based approach. Available: PMM (default),
CBMI}
}
\value{
Point estimation and variance estimation obtained by delta method:
\itemize{
\item{llh_est}{point estimate}
\item{var_llh}{variance estimate approximated by delta method}
}
}
\description{
For a given dataset with monotone missingness, find the likelihood-based
inference of the ATE
}
