% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{j2r_seq_llh}
\alias{j2r_seq_llh}
\title{Sequential likelihood-based inference from delta method}
\usage{
j2r_seq_llh(
  seq_formula,
  data_wide,
  trt_name,
  M = 10,
  fit_model,
  method = "PMM"
)
}
\arguments{
\item{seq_formula}{a list of formulas for sequential regression}

\item{data_wide}{wide-form of the data}

\item{trt_name}{name of the treatment variable}

\item{M}{imputation size}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regression), MASS::rlm (robust regression)}

\item{method}{name of the likelihood-based method. Available: PMM, CBMI}
}
\value{
Point estimation and variance estimation obtained by delta method:
\itemize{
\item{llh_est}{point estimate}
\item{var_llh}{variance estimate approximated by delta method}
}
}
\description{
Compute variance estimate of the likelihood-based inference using delta
method
}
