% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{j2rLlh}
\alias{j2rLlh}
\title{Likelihood-based method}
\usage{
j2rLlh(
  mmrm_formula,
  data_long,
  time_name,
  id_name,
  trt_name,
  method = "PMM",
  bootstrap = FALSE,
  B = 100
)
}
\arguments{
\item{mmrm_formula}{formula for fixed effect MMRM under MAR}

\item{data_long}{long-form of the data frame}

\item{time_name}{name of the time variable}

\item{id_name}{name of the id variable}

\item{trt_name}{name of the treatment variable}

\item{method}{specific method for likelihood-based approach. Available: PMM (default),
CBMI}

\item{bootstrap}{whether to use bootstrap variance}

\item{B}{the number of bootstrap replicate. Default: B = 100}
}
\value{
Point estimation and variance estimation obtained by delta method:
\itemize{
\item{ate_est}{point estimate}
\item{var_llh}{variance estimate approximated by delta method}
\item{var_boot}{variance estimate approximated by delta method}
}
}
\description{
For a given dataset, find the likelihood-based inference of the ATE by
delta method or nonparametric bootstrap
}
