% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kr_comp.R
\name{kr}
\alias{kr}
\title{obtain Kenward-Roger degree of freedom}
\usage{
kr(fit, contrast, order = 2L)
}
\arguments{
\item{fit}{\code{mmrm} fit}

\item{contrast}{contrast vector (ncol = 1)}
}
\description{
obtain Kenward-Roger degree of freedom
}
\examples{
fit <- mmrm(FEV1 ~ ARMCD + ar1(AVISIT | USUBJID), data = fev_data)
contrast <- matrix(c(0,1), ncol = 1)
kr(fit, contrast)
}
