% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\alias{h_is_specified}
\alias{h_is_specified_and_in_data}
\alias{h_check_and_get_label}
\alias{h_labels}
\title{Adding Labels To Variables For Model}
\usage{
h_is_specified(x, vars)

h_is_specified_and_in_data(x, vars, data)

h_check_and_get_label(x, vars, data)

h_labels(vars, data)
}
\arguments{
\item{x}{(\code{character})\cr an element in vars.}

\item{vars}{(\code{list})\cr variables to use.}

\item{data}{(\code{data.frame})\cr data to use.}
}
\value{
list of variables with labels
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Functions}{
\itemize{
\item \code{h_is_specified}: checks if element in \code{vars} is not \code{NULL}.

\item \code{h_is_specified_and_in_data}: checks if element in vars is not NULL and exists in dataset.

\item \code{h_check_and_get_label}: gets label for each element in vars.

\item \code{h_labels}: returns list of variables with labels.
}}

\examples{
data <- data.frame(
  MYID = c(1, 1, 2, 2, 3, 3),
  ARM = factor(c(1, 2, 3, 1, 2, 3)),
  AVAL = c(2, 4, 6, 8, 10, 12),
  AVISIT = c("W1", "W2", "W3", "W1", "W2", "W3"),
  SEX = c("Female", "Female", "Female", "Male", "Male", "Male"),
  RACE = c("Asian", "Asian", "Asian", "White", "Asian", "White")
)
vars <- list(response = "AVAL", id = "MYID", arm = "ARM", visit = "AVISIT", covariates = c("SEX", "RACE"))
h_labels(vars, data)
}
