% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{linearHypothesis.mmrm}
\alias{linearHypothesis.mmrm}
\title{Test a Linear Hypothesis on an \code{mmrm} Fit}
\usage{
linearHypothesis.mmrm(
  model,
  hypothesis.matrix,
  rhs = NULL,
  vcov. = NULL,
  singular.ok = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model}{(\code{mmrm})\cr an \code{mmrm} model fit.}

\item{hypothesis.matrix}{(\code{matrix})\cr see \code{\link[car:linearHypothesis]{car::linearHypothesis()}}.}

\item{rhs}{(\code{numeric})\cr see \code{\link[car:linearHypothesis]{car::linearHypothesis()}}.}

\item{vcov.}{(\code{matrix} or \code{function})\cr the coefficient-covariance matrix or
a function to compute a covariance matrix. This is only used when
\code{test.statistic = "Chisq"}.}

\item{singular.ok}{(\code{flag})\cr flag indicating whether or not to allow models
with aliased coefficients to undergo \code{Chisq} testing. Only used if
\code{test.statistic = "Chisq"}.}

\item{verbose}{(\code{flag})\cr see \code{\link[car:linearHypothesis]{car::linearHypothesis()}}.}

\item{...}{arguments passed down from other methods.}
}
\description{
Test a Linear Hypothesis on an \code{mmrm} Fit
}
\keyword{internal}
