% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-package.R
\docType{package}
\name{mmrm-package}
\alias{mmrm-package}
\title{\code{mmrm} Package}
\description{
\code{mmrm} implements mixed models with repeated measures (MMRM) in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openpharma.github.io/mmrm/}
  \item Report bugs at \url{https://github.com/openpharma/mmrm/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanes Bove \email{daniel.sabanes_bove@rconis.com} (\href{https://orcid.org/0000-0002-0176-9239}{ORCID})

Authors:
\itemize{
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@rconis.com} (\href{https://orcid.org/0000-0002-0176-9239}{ORCID})
  \item Liming Li \email{liming.li1@astrazeneca.com} (\href{https://orcid.org/0009-0008-6870-0878}{ORCID})
  \item Julia Dedic \email{julia.dedic@roche.com}
  \item Doug Kelkhoff \email{doug.kelkhoff@roche.com}
  \item Kevin Kunzmann \email{kevin.kunzmann@boehringer-ingelheim.com}
  \item Brian Matthew Lang \email{brian.lang@msd.com}
  \item Christian Stock \email{christian.stock@boehringer-ingelheim.com}
  \item Ya Wang \email{ya.wang10@gilead.com}
  \item Dan James \email{dan.james@astrazeneca.com}
  \item Jonathan Sidi \email{yoni@pinpointstrategies.io}
  \item Daniel Leibovitz \email{daniel.leibovitz@roche.com}
  \item Daniel D. Sjoberg \email{sjobergd@gene.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
  \item Nikolas Ivan Krieger \email{nikolas.krieger@experis.com} (\href{https://orcid.org/0000-0002-4581-3545}{ORCID})
}

Other contributors:
\itemize{
  \item Craig Gower-Page \email{craig.gower-page@roche.com} [contributor]
  \item Lukas A. Widmer (\href{https://orcid.org/0000-0003-1471-3493}{ORCID}) [contributor]
  \item Boehringer Ingelheim Ltd. [copyright holder, funder]
  \item Gilead Sciences, Inc. [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Merck Sharp & Dohme, Inc. [copyright holder, funder]
  \item AstraZeneca plc [copyright holder, funder]
  \item inferential.biostatistics GmbH [copyright holder, funder]
}

}
