% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_cv.R
\name{mmrm_cv}
\alias{mmrm_cv}
\title{K-fold cross validation of MMRM models}
\usage{
mmrm_cv(
  formula,
  time,
  subjects,
  data,
  k = 10,
  seed = NULL,
  in_loop = NULL,
  ...
)
}
\arguments{
\item{formula}{formula for the fixed effects structure of the model}

\item{time}{the time variable of the model}

\item{subjects}{the variable indicating unique subjects}

\item{data}{the data structure}

\item{k}{the number of splits for k-fold cross-validation, default = 10}

\item{seed}{random seed used to generating splits,
If NULL does not set the seed}

\item{in_loop}{function to process each split of training set data. See details}

\item{...}{arguments passed to \code{in_loop} and \link{MMRM}}
}
\value{
\code{mmrmCV} object: a list of outputs with length \code{k},
each element is the output of a call to \link{mmrm}.
See \link{mmrm} and \link{mmrmObject} for details
}
\description{
Performs k-fold cross validation with the Mixed Model Repeated Measures
model (see \link{MMRM}).
}
\details{
Please see \link{MMRM} for details on the model.

\link{mmrm_cv} supports fitting each of the k-folds in parallel using
\link[foreach:foreach]{foreach::foreach} loops. To use multiple cores, please register a parallel
backend prior to calling \link{mmrm_cv}. Here is an example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{cl = parallel::makeCluster(n_cores)
doParallel::registerDoParallel(cl)
}\if{html}{\out{</div>}}

The \code{in_loop} argument allows users to provide a function that
transforms training set data. This function must:
\itemize{
\item accept the training set data as its first argument
\item accept pass through arguments (...)
\item return the processed data structure with all variables
specified in the supplied formula
Here is an example \code{in_loop} that performs no transformations:
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{in_loop = function(data, ...) \{
  # do stuff here
  return(data)
\}
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
mmrm_cv(
  outcome ~ baseline + group + time +
    baseline:time + group:time,
  time = "time",
  subjects = "subjectid",
  data = my_data,
  k = 10
)
}

}
