% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_eff_size.R
\name{mmrm_eff_size}
\alias{mmrm_eff_size}
\title{Calculate effect size for contrasts using estimated marginal means}
\usage{
mmrm_eff_size(mmrm_model, mmrm_emm, ...)
}
\arguments{
\item{mmrm_model}{the mmrm model(s) of type \code{mmrm}, \code{mmrmList}, or \code{mmrmCV}
used to calculate effect size}

\item{mmrm_emm}{estimated marginal means from and MMRM.
output of \link{mmrm_emmeans}}

\item{...}{additional arguments passed to \link[emmeans:eff_size]{emmeans::eff_size}}
}
\description{
Wrapper for \link[emmeans:eff_size]{emmeans::eff_size} to run for MMRM models.
}
\details{
The \code{sigma} argument (the population standard deviation) for
\link[emmeans:eff_size]{emmeans::eff_size} is calculated using \link[stats:sigma]{stats::sigma}. The \code{edf} argument
(the equivalent degrees of freedom) is estimated as the mean of the
degrees of freedom  across all contrasts calculated in \link{mmrm_emmeans}
(typically using Kenward-Rogers approximation).
}
