% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_emmeans.R
\name{mmrm_emmeans}
\alias{mmrm_emmeans}
\title{Estimated Marignal Means of MMRM models}
\usage{
mmrm_emmeans(obj, ...)
}
\value{
see \link[emmeans:emmeans]{emmeans::emmeans}
}
\description{
Generic method to run \link[emmeans:emmeans]{emmeans::emmeans} on all \link{MMRM} package objects using
Kenward-Rogers degrees of freedom calculation.
Please see \link[emmeans:emmeans]{emmeans::emmeans} for further documentation.
}
\details{
Support for Kenward-Rogers degrees of freedom calculation
is made possible with i) a custom emm_basis function and
ii) addition of support for glsObject in the pbkrtest package
(currently in a development version
\href{https://github.com/gkane26/pbkrtest/tree/nlme}{gkane26/pbkrtest@nlme}).
\cr\cr
Kenward-Rogers d.f. calculation is memory and computation time intensive.
By default, running this method on a large dataset will throw an error to warn users.
To bypass this error, please set the pbkrtest.limit argument to a number larger than
the number of observations in the model.
}
\examples{
\dontrun{
mmrm_emmeans(
  object = mmrm_object,
  specs = pairwise ~ time | treatment,
  mode = "kenward",
  pbkrtest.limit = emmeans::get_emm_option("pbkrtest.limit")
)
}

}
