% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-methods.R
\name{mmrm_methods}
\alias{mmrm_methods}
\alias{summary.mmrm}
\alias{print.summary.mmrm}
\title{Methods for \code{mmrm} Objects}
\usage{
\method{summary}{mmrm}(object, ...)

\method{print}{summary.mmrm}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{(\code{mmrm})\cr the fitted MMRM including Jacobian and call etc.}

\item{...}{not used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Functions}{
\itemize{
\item \code{summary.mmrm}: summarizes the MMRM fit results.

\item \code{print.summary.mmrm}: prints the MMRM fit summary.
}}

\examples{
formula <- FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
object <- mmrm(formula, fev_data)
# Summary:
summary(object)
}
\keyword{internal}
