% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_simulate.R
\name{mmrm_simulate}
\alias{mmrm_simulate}
\title{Simulate MMRM data}
\usage{
mmrm_simulate(
  n_per_arm = 50,
  n_timepoints = 5,
  effect_tp = 8,
  effect_baseline = -0.5,
  effect_arm = 2,
  covariance = NULL,
  variance = 10,
  correlation = 0,
  baseline = c(25, 5),
  p_missing = 0.1
)
}
\arguments{
\item{n_per_arm}{the number of subjects per arm}

\item{n_timepoints}{the number of timepoints.
If effects are specified as vectors, this is ignored.}

\item{effect_tp}{mean observation at each timepoint.
Either a vector with length equal to the
number of timepoints or a scalar.}

\item{effect_baseline}{effect of baseline value at each timepoint.
Either a vector with length equal to the
number of timepoints or a scalar.}

\item{effect_arm}{effect of arm at each timepoint.
Either a vector with length equal to the
number of timepoints or a scalar.}

\item{covariance}{full residual covariance matrix.
Either covariance of both of variance and
correlation should be specified.}

\item{variance}{variance at each timepoint.
Either a vector with length equal to the
number of timepoints or a scalar.
Either both of variance and correlation or
covariance should be specified}

\item{correlation}{correlation of residuals across timepoints.
Should be a scalar or vector with length:
timepoints * (timepoints - 1) / 2.
Either both of variance and correlation or
covariance should be specified.}

\item{baseline}{mean and variance of baseline scores}

\item{p_missing}{probability of missing data at each timepoint.
Either a vector with length equal to the
number of timepoints or a scalar.}
}
\value{
data.frame with simulated data in the long format
with the following fields:
\itemize{
\item n_subs: the nubmer of subjects
\item subject: subject id
\item arm: the study arm or group
\item base: baseline score
\item time: time point of observation
\item chg: the change score at the given time point (i.e., the outcome)
}
}
\description{
Simulate a "clinical" dataset according to assumptions of the MMRM.
Datasets should have "observations" at a selected number of discrete timepoints.
Observations are predicted by:
\itemize{
\item the mean observation at each timepoint
\item the effect of baseline at each timepoint
\item the effect of arm at each timepoint
\item residual covariance matrix
}
}
