% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{residuals.mmrm_tmb}
\alias{residuals.mmrm_tmb}
\title{Residuals method for mmrm objects}
\usage{
\method{residuals}{mmrm_tmb}(object, type = c("response", "pearson", "normalized"))
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM.}

\item{type}{(\code{character})\cr unscaled ('response') or 'normalized'.}
}
\value{
Vector of residuals
}
\description{
Residuals method for mmrm objects
}
\section{Functions}{
\itemize{
\item \code{residuals(mmrm_tmb)}: to obtain residuals - either unscaled or normalized

}}
\examples{
# residuals:
residuals(object, type = "response")
}
