% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{mmrm_tmb_methods}
\alias{mmrm_tmb_methods}
\alias{logLik.mmrm_tmb}
\alias{formula.mmrm_tmb}
\title{Methods for \code{mmrm_tmb} Objects}
\usage{
\method{logLik}{mmrm_tmb}(object, ...)

\method{formula}{mmrm_tmb}(x, ...)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM object.}

\item{...}{not used.}

\item{x}{(\code{mmrm_tmb})\cr same as \code{object}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Functions}{
\itemize{
\item \code{logLik.mmrm_tmb}: obtains the attained log likelihood value.

\item \code{formula.mmrm_tmb}: obtains the used formula.
}}

\examples{
formula <- FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
data <- fev_data
object <- h_mmrm_tmb(formula, data)
# Log likelihood given the estimated parameters:
logLik(object)
# Formula which was used:
formula(object)
}
