% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{mmrmformula}
\alias{mmrmformula}
\title{Create formula that fits the MMRM model (fixed + random effect)}
\usage{
mmrmformula(mmrm_formula, time_name, id_name)
}
\arguments{
\item{mmrm_formula}{formula for fixed effect in MMRM}

\item{time_name}{name of the time variable}

\item{id_name}{name of the id variable}
}
\description{
Create formula that fits the MMRM model (fixed + random effect)
}
