% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_object.R
\name{mmrmObject}
\alias{mmrmObject}
\title{Fitted mmrm Object}
\value{
"mmrm" objects have all components of \link[nlme]{glsObject},
plus the following:
\item{data}{the data structure used to fit the model}
\item{warnings}{warning messages from the model fitting procedure}
}
\description{
An object returned by the \link{mmrm} function,
representing a fitted Mixed Model Repeated Measures model.
mmrm objects inherit from class "mmrm" and "gls", abd are a simple
extension of \link[nlme:glsObject]{nlme::glsObject}.
Please see \link[nlme:glsObject]{nlme::glsObject} for full details.
}
\seealso{
\link[nlme:glsObject]{nlme::glsObject} \code{\link[=mmrm]{mmrm()}}
}
