% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{naive_est}
\alias{naive_est}
\title{Point estimation using naive method, under J2R}
\usage{
naive_est(
  mmrm_formula,
  data_long,
  time_name,
  id_name,
  trt_name,
  M = 10,
  fit_model,
  estimator = "reg with interaction"
)
}
\arguments{
\item{mmrm_formula}{a symbolic description of the model to be fitted}

\item{data_long}{long-form of the data frame}

\item{time_name}{name of the time variable}

\item{id_name}{name of the id variable}

\item{trt_name}{name of the treatment variable}

\item{M}{imputation size}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regression), MASS::rlm (robust regression)}

\item{estimator}{Estimator type in Rubin's method}
}
\value{
Estimation results using naive method.
\itemize{
\item{ate_est}{MI point estimation of the ATE}
\item{var_rubin}{MI variance estimation by Rubin's rule}
}
}
\description{
Obtain estimation of the ATE using naive method by Rubin's rule
}
