% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{nonpara_fn}
\alias{nonpara_fn}
\title{Get bootstrap variance estimate}
\usage{
nonpara_fn(
  formula,
  data_wide = NULL,
  data_long = NULL,
  time_name = NULL,
  id_name,
  trt_name,
  M,
  B,
  fit_model,
  estimator,
  method,
  llh_method = NULL
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{data_wide}{wide-form of the data frame}

\item{data_long}{long-form of the data frame}

\item{time_name}{name of the time variable (needed for long-form data)}

\item{id_name}{name of the id variable}

\item{trt_name}{name of the treatment variable}

\item{M}{imputation size}

\item{B}{number of bootstrap replicates}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regressoin), MASS::rlm (robust regression)}

\item{estimator}{type of the estimator. Available: sample mean,
reg with interaction, reg without interaction}

\item{method}{several methods to use. Available: naive MI, sequential MI,
naive likelihood, sequential likelihood}

\item{llh_method}{name of the likelihood-based method. Available: PMM, CBMI}
}
\value{
variance estimation by nonparametric bootstrap
}
\description{
For a given dataset, compute the bootstrap variance estimate by a
prespecified method
}
