% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-parsnip.R
\name{parsnip_add_mmrm}
\alias{parsnip_add_mmrm}
\title{Register mmrm for use with tidymodels}
\usage{
parsnip_add_mmrm(quietly = FALSE)
}
\arguments{
\item{quietly}{logical: should progress and error messages be suppressed?}
}
\value{
A logical value indicating whether registration was successful
}
\description{
Register mmrm for use with tidymodels
}
\keyword{internal}
