% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plumber_api}
\alias{plumber_api}
\title{Personal greeting as a Plumber API}
\usage{
plumber_api(...)
}
\arguments{
\item{...}{Additional arguments to plumber::pr_run()}
}
\value{
Plumber API showcasing the personal greeting feature.
}
\description{
Greet a person and appropriately capitalize their name
as a Plumber API.
}
