% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{position_symbol}
\alias{position_symbol}
\title{Search For the Position of a Symbol}
\usage{
position_symbol(x, sym, ...)
}
\arguments{
\item{x}{(\code{list} of \code{language})\cr a list of language objects in which to
search for a specific symbol.}

\item{sym}{(\code{name} or \code{symbol} or \code{character})\cr a symbol to search for in
\code{x}.}

\item{...}{Additional arguments passed to \code{Position()}.}
}
\value{
The position of the symbol if found, or the \code{nomatch} value
otherwise.
}
\description{
A thin wrapper around \code{\link[base:funprog]{base::Position()}} to search through a list of language
objects, as produced by \code{\link[=flatten_call]{flatten_call()}} or \code{\link[=flatten_expr]{flatten_expr()}}, for the
presence of a specific symbol.
}
\keyword{internal}
