% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{print.cov_struct}
\alias{print.cov_struct}
\title{Print a Covariance Structure Object}
\usage{
\method{print}{cov_struct}(x, ...)
}
\arguments{
\item{x}{(\code{cov_struct})\cr a covariance structure object.}

\item{...}{Additional arguments unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a Covariance Structure Object
}
