% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{register_on_load}
\alias{register_on_load}
\title{Helper function for registering functionality with suggests packages}
\usage{
register_on_load(pkg, ver = c(NA, NA), callback, message = NULL)
}
\arguments{
\item{pkg}{A package name}

\item{ver}{A vector of length two, specifying minimum and maximum (inclusive)
versions for interoperability.}

\item{message}{An optional message to print if mmrm functionality is
successfully registered}
}
\value{
A logical (invisibly) indicating whether registration was successful.
If not, a onLoad hook was set for the next time the package is loaded.
}
\description{
Helper function for registering functionality with suggests packages
}
\keyword{ineternal}
