% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{rr_seq_imp}
\alias{rr_seq_imp}
\title{Sequential imputation}
\usage{
rr_seq_imp(
  seq_formula,
  data_wide,
  trt_name,
  M,
  fit_model,
  estimator = "reg with interaction"
)
}
\arguments{
\item{seq_formula}{a list of formulas for sequential regression}

\item{data_wide}{wide-form of the data frame}

\item{trt_name}{name of the treatment variable}

\item{M}{imputation size}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regression), MASS::rlm (robust regression)}

\item{estimator}{type of the estimator. Available: sample mean,
reg with interaction, reg without interaction}
}
\value{
Rubin's point and variance estimate
\itemize{
\item{ate_est}{MI point estimation of the ATE}
\item{var_rubin}{MI variance estimation by Rubin's rule}
}
}
\description{
For a given dataset with monotone missingness, conduct sequential imputation
based on linear / robust / rank regression
}
