% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{rubin_est}
\alias{rubin_est}
\title{Get Rubin's estimate}
\usage{
rubin_est(chg_imp, data_wide, covar_name, trt_name, M, fit_model, estimator)
}
\arguments{
\item{chg_imp}{a matrix of imputed response at the last time point}

\item{data_wide}{wide-form of the data frame}

\item{covar_name}{names of the baseline covariates}

\item{trt_name}{name of the treatment variable}

\item{M}{imputation size}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regressoin), MASS::rlm (robust regression)}

\item{estimator}{type of the estimator. Available: sample mean,
reg with interaction, reg without interaction. Default: reg with interaction.}
}
\value{
Rubin's point and variance estimate
\itemize{
\item{ate_est}{MI point estimation of the ATE}
\item{var_rubin}{MI variance estimation by Rubin's rule}
}
}
\description{
Get estimation results by Rubin's rule
}
