% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{seq_mi_fit}
\alias{seq_mi_fit}
\title{MI-based Sequential regression}
\usage{
seq_mi_fit(seq_formula, data_wide, trt_name, M, fit_model)
}
\arguments{
\item{seq_formula}{a list of formulas for sequential regression}

\item{data_wide}{wide form of the longitudinal data}

\item{trt_name}{name of treatment in the data frame}

\item{M}{imputation size}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regression), MASS::rlm (robust regression)}
}
\value{
A list of estimation results from sequential regression:
\itemize{
\item{obs_pi}{marginal probability of the observed data for each treatment.}
\item{beta_ctl_imp}{M sets of the estimated MMRM coefficients used in MI
for control group.}
\item{beta_trt_imp}{M sets of the estimated MMRM coefficients used in MI
for treatment group.}
\item{sigma_ctl}{Estimated population covariance matrix of the longitudinal
response.}
\item{beta_rubin_ctl}{Estimated coefficients at the lsat time point for
control group.}
\item{beta_rubin_trt}{Estimated coefficients at the lsat time point for
treatment group.}
\item{var_rubin_ctl}{Estimated covariance matrix of the coefficients at
the lsat time point for control group.}
\item{var_rubin_trt}{Estimated covariance matrix of the coefficients at
the lsat time point for treatment group.}
}
}
\description{
Conduct sequential imputation under MAR to get parameter estimates for MMRM.
}
