% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{seqLlh}
\alias{seqLlh}
\title{Inference using MI-based sequential based method, under J2R}
\usage{
seqLlh(
  seq_formula,
  data_wide,
  trt_name,
  M,
  fit_model,
  id_name,
  method = "PMM",
  bootstrap = FALSE,
  B = 100
)
}
\arguments{
\item{seq_formula}{a list of formulas for sequential regression}

\item{data_wide}{wide-form of the data frame}

\item{trt_name}{name of the treatment variable}

\item{M}{imputation size}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regression), MASS::rlm (robust regression)}

\item{id_name}{name of the id variable}

\item{method}{name of the likelihood-based method. Available: PMM, CBMI}

\item{bootstrap}{Whether to use bootstrap variance}

\item{B}{the number of bootstrap replicate}
}
\value{
Point estimation and variance estimation obtained by delta method:
\itemize{
\item{llh_est}{point estimate}
\item{var_llh}{variance estimate approximated by delta method}
\item{var_boot}{variance estimate approximated by delta method}
}
}
\description{
For a given dataset with monotone missingness, impute the missing data from
the conditional distribution given observed under J2R by proper MI, output
the imputed dataset
}
