% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{seqMI}
\alias{seqMI}
\title{Inference using MI-based sequential based method, under J2R}
\usage{
seqMI(
  seq_formula,
  data_wide,
  trt_name,
  M,
  fit_model,
  id_name,
  estimator = "reg with interaction",
  bootstrap = FALSE,
  B = 100
)
}
\arguments{
\item{seq_formula}{a list of formulas for sequential regression}

\item{data_wide}{wide-form of the data frame}

\item{trt_name}{name of the treatment variable}

\item{M}{imputation size}

\item{fit_model}{type of the analysis model. Available: lm, Rfit::rfit
(rank regression), MASS::rlm (robust regression)}

\item{id_name}{name of the id variable}

\item{estimator}{type of the estimator. Available: sample mean,
reg with interaction, reg without interaction. Default: reg with interaction}

\item{bootstrap}{Whether to use bootstrap variance}

\item{B}{the number of bootstrap replicate}
}
\value{
A list of MI estimation of the ATE
\itemize{
\item{ate_est}{MI point estimation of the ATE}
\item{var_rubin}{MI variance estimation by Rubin's rule}
\item{var_boot}{MI variance estimation by nonparametric bootstrap}
}
}
\description{
Given a wide-form data, compute the estimate of the ATE under J2R using
MI-based sequential method.
}
