% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{sim_one_group}
\alias{sim_one_group}
\title{Simulate longitudinal data with monotone missingness for one group}
\usage{
sim_one_group(
  n,
  k,
  P,
  sigma,
  beta,
  pi = NULL,
  pi_par = NULL,
  trt = 1,
  missing_type,
  distribution,
  outlier = "none"
)
}
\arguments{
\item{n}{sample size}

\item{k}{dimension of the covariates}

\item{P}{the number of time point}

\item{sigma}{Covariance matrix for the longitudinal response}

\item{beta}{regression coefficient for the baseline covariates}

\item{pi}{Observed probability for each missing pattern under MCAR}

\item{pi_par}{Parameters for the missing mechanism model under MAR}

\item{trt}{name of the treatment}

\item{missing_type}{Type of missing. Available: MCAR, MAR, MNAR}

\item{distribution}{Underlying multivariate distribution for the longitudinal responses}

\item{outlier}{The types of outlier may exists. Available: none, Type 1, Type 2}
}
\value{
a list including a wide-form data frame and a long-form data frame
\itemize{
\item{db_comb}{Wide form of the original longitudinal data}
\item{db_long}{Long form of the original longitudinal data}
\item{db_comb0}{Wide form of the change from baseline longitudinal data}
\item{db_long0}{Long form of the change from baseline longitudinal data}
}
}
\description{
Generate longitudinal data with monotone missingness
}
