% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{start_funs}
\alias{start_funs}
\title{Starting Value Functions}
\format{
An object of class \code{list} of length 3.
}
\usage{
start_funs
}
\description{
An internal constant of available start value functions.
}
\keyword{internal}
