% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_start}
\alias{std_start}
\title{Standard Starting Value}
\usage{
std_start(cov_type, n_visits, n_groups, ...)
}
\arguments{
\item{cov_type}{(\code{string})\cr name of the covariance structure.}

\item{n_visits}{(\code{int})\cr number of visits.}

\item{n_groups}{(\code{int})\cr number of groups.}

\item{...}{not used.}
}
\value{
A numeric vector of starting values.
}
\description{
Obtain standard start values.
}
\keyword{internal}
