% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{summaryLong}
\alias{summaryLong}
\title{Summary of long-form data frame given sequential regression formula}
\usage{
summaryLong(mmrm_formula, data_long, time_name, id_name, trt_name)
}
\arguments{
\item{mmrm_formula}{formula for fixed effect in MMRM}

\item{data_long}{long form of the data}

\item{time_name}{name of the time variable}

\item{id_name}{name of the id variable}

\item{trt_name}{name of the treatment variable}
}
\value{
A list of summary of the long form data:
\itemize{
\item{p} number of time points
\item{covar_name}{name of the covariates}
\item{y_name}{name of the response}
\item{k}{dimension of the covariates}
\item{adj_mean}{overall covariate mean}
\item{var_xbar}{covariance matrix of the covariates}
\item{obs_pi}{marginal probability of the observed data for each treatment}
\item{db_avaiable_ctl}{long-form available data in the control group}
\item{db_avaiable_trt}{long-form available data in the treatment group}
\item{data_wide}{wide-form data}
\item{data_wide_ctl}{wide-form data in the control group}
\item{data_wide_trt}{wide-form data in the treatment group}
\item{n1}{number of subjects in the control group}
\item{n2}{number of subjects in the treatment group}
\item{data_complete_ctl}{complete data in the control group}
\item{data_complete_trt}{complete data in the treatment group}
}
}
\description{
Given a long-form data and a formula from MMRM, summarize the information
from the data
}
