% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{summaryWide}
\alias{summaryWide}
\title{Summary of wide-form data frame given sequential regression formula}
\usage{
summaryWide(seq_formula, data_wide, trt_name)
}
\arguments{
\item{seq_formula}{a list of formulas for sequential regression}

\item{data_wide}{wide-form of the data}

\item{trt_name}{name of the treatment variable}
}
\value{
A list of summary of the long form data:
\itemize{
\item{p} number of time points
\item{covar_name}{name of the covariates}
\item{k}{dimension of the covariates}
\item{last_name}{name of the response at last time point}
\item{adj_mean}{overall covariate mean}
\item{var_xbar}{covariance matrix of the covariates}
\item{obs_pi}{marginal probability of the observed data for each treatment}
\item{n1}{number of subjects in the control group}
\item{n2}{number of subjects in the treatment group}
\item{data_complete_ctl}{complete data in the control group}
\item{data_complete_trt}{complete data in the treatment group}
}
}
\description{
Summary of wide-form data frame given sequential regression formula
}
