% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.mmrm}
\alias{tidy.mmrm}
\alias{glance.mmrm}
\alias{augment.mmrm}
\title{Tidying methods for an mmrm model}
\usage{
\method{tidy}{mmrm}(x, conf.int = FALSE, conf.level = 0.95)

\method{glance}{mmrm}(x)

\method{augment}{mmrm}(
  x,
  newdata = NULL,
  interval = c("none", "confidence", "prediction"),
  se_fit = interval != "none",
  type.residuals = c("response", "pearson", "normalized"),
  ...
)
}
\arguments{
\item{x}{Fitted model object from the mmrm package}

\item{conf.int}{If TRUE columns for the lower (conf.low) and upper bounds
(conf.high) of coefficient estimates are included., Default: FALSE}

\item{conf.level}{Defines the range of the confidence internal.
Only used if conf.int = TRUE, Default: 0.95}

\item{newdata}{data.frame, new data frame, Default: NULL}

\item{interval}{type of interval calculation. Can be abbreviated,
Default: c("none", "confidence", "prediction")}

\item{se_fit}{logical, return standard errors of fit,
Default: interval!='none'}

\item{type.residuals}{Passed on to residuals.mmrm_tmb().
Default: c("response", "pearson", "normalized")}

\item{\dots}{arguments passed on to predict method}
}
\value{
All tidying methods return a tibble.
The structure depends on the method chosen.
}
\description{
These methods tidy the estimates from an mmrm_tmb object into a
summary
}
\examples{
if(interactive()){

  # fit a model
    fit <- mmrm(FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID), data = fev_data)

  # tidy method
    fit |> tidy()
    fit |> tidy(conf.int = TRUE)
    fit |> tidy(conf.int = TRUE, conf.level = 0.9)

  # glance method
    fit |> glance()

  # augment method
    fit |> augment()
    fit |> augment(interval = 'confidence')
    fit |> augment(interval = 'prediction')

    fit |> augment(type.residuals='pearson')
    fit |> augment(type.residuals='normalized')
 }
}
\seealso{
\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{enframe}}
}
