% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{tmb_cov_type}
\alias{tmb_cov_type}
\title{Produce A Covariance Identifier Passing to TMB}
\usage{
tmb_cov_type(cov)
}
\arguments{
\item{cov}{(\code{cov_struct})\cr a covariance structure object.}
}
\value{
A string used for method dispatch when passed to TMB.
}
\description{
Produce A Covariance Identifier Passing to TMB
}
\keyword{internal}
