% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{transform_coef}
\alias{transform_coef}
\title{Transform the regression coefficients for lmer() object}
\usage{
transform_coef(x, p, k)
}
\arguments{
\item{x}{regression coefficient returns from lmer()}

\item{p}{the number of time point}

\item{k}{dimension of the covariates}
}
\value{
transformed estimated coefficients at each time point
}
\description{
Transform the regression coefficients for lmer() object
}
