% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fn.R
\name{transform_cov}
\alias{transform_cov}
\title{Transform the population covariance matrix for lmer() object}
\usage{
transform_cov(x, p)
}
\arguments{
\item{x}{a data frame of the lower triangle elements obtained from lme4::VarCorr()}

\item{p}{the number of time point}
}
\value{
Estimated population covariance matrix
}
\description{
Transform the population covariance matrix for lmer() object
}
