% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{typed_param}
\alias{typed_param}
\title{Format 'roxygen2' Type-Annotated Parameter Definitions}
\usage{
typed_param(name, type, description)
}
\arguments{
\item{name}{(\code{symbol | character(1L)})\cr A parameter name to annotate.}

\item{type}{(\code{expr | character(1L)})\cr A type signature to use for annotation.}

\item{description}{(\code{character(1L)})\cr Parameter description.}
}
\value{
A 'roxygen2'-formatted \verb{@param} string.
}
\description{
Format 'roxygen2' Type-Annotated Parameter Definitions
}
\keyword{internal}
