% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{validate_cov_struct}
\alias{validate_cov_struct}
\title{Validate Covariance Structure Data}
\usage{
validate_cov_struct(x)
}
\arguments{
\item{x}{(\code{cov_struct})\cr a covariance structure object.}
}
\value{
\code{x} if successful, or an error is thrown otherwise.
}
\description{
Run checks against relational integrity of covariance definition
}
\keyword{internal}
