% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcov.R
\name{varcov}
\alias{varcov}
\title{Get Variance-Covariance matrix of MMRM model}
\usage{
varcov(model)
}
\arguments{
\item{model}{a glsObject (output of \link[MMRM]{mmrm})}
}
\value{
matrix, the variance-covoariance matrix
}
\description{
Copied from \link[SPR]{VarCor_gls}
}
