% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checkmate.R
\name{with_error_call}
\alias{with_error_call}
\title{Replace Errors's Call Before Raising}
\usage{
with_error_call(expr, call = sys.call(-1))
}
\arguments{
\item{expr}{(\code{expression})\cr
An expression to evaluate in the parent frame, which may raise a
\code{condition}.}

\item{call}{(\code{call})\cr
A call to associate with any errors before they are raised.}
}
\value{
The result of evaluating \code{expr} if no errors are raised. An error is
raised is otherwise.
}
\description{
To circumvent \code{checkmate::reportAssertions}' immediate raising of errors with
a top-level call, this function captures the errors and applies a call to the
condition before it is raised, providing more localized and informative error
messages.
}
\keyword{internal}
