% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{Anova.mmrm}
\alias{Anova.mmrm}
\title{Conduct type II/III hypothesis testing on the MMRM fit results.}
\usage{
Anova.mmrm(
  mod,
  type = c("II", "III", "2", "3"),
  tol = sqrt(.Machine$double.eps),
  test.statistic = c("F", "Chisq"),
  vcov. = vcov(mod, complete = FALSE),
  ...
)
}
\arguments{
\item{mod}{(\code{mmrm})\cr the fitted MMRM.}

\item{type}{(\code{string}) type of test, "II", "III", '2', or '3'.}

\item{tol}{(\code{numeric}) threshold blow which values are treated as 0.}

\item{test.statistic}{(\code{string})\cr either \verb{"F} or \code{"Chisq"}, indicating the
kind of test to perform.}

\item{vcov.}{(\code{matrix} or \code{function})\cr the coefficient-covariance matrix or
a function to compute a covariance matrix. This is only used when
\code{test.statistic = "Chisq"}.}

\item{...}{Passed to \code{\link[car:linearHypothesis]{car::linearHypothesis()}} when \code{test.statistic = "Chisq"}.}

\item{singular.ok}{(\code{flag})\cr flag indicating whether or not to allow models
with aliased coefficients to undergo \code{Chisq} testing. Only used if
\code{test.statistic = "Chisq"}.}
}
\description{
Conduct type II/III hypothesis testing on the MMRM fit results.
}
\details{
\code{Anova()} will return \code{anova} object with one row per variable. If
\code{test.statistic = "F"}, columns will be \verb{Num Df}(numerator degrees of
freedom), \verb{Denom Df} (denominator degrees of freedom), \verb{F Statistic} and
\verb{Pr(>=F)}.

If \code{test.statistic = "Chisq"}, columns will be \code{Chisq} (the Chi-squared
test statistic), \code{Df} (degrees of freedom), and \verb{Pr(>Chisq)} (p-value).
}
\examples{

\dontshow{

Please do not load `car` and then create the documentation. The Rd file will
be different.}
}
\keyword{internal}
