% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-car.R
\name{Anova.mmrm}
\alias{Anova.mmrm}
\title{Conduct type II/III hypothesis testing on the MMRM fit results.}
\usage{
Anova.mmrm(
  mod,
  type = c("II", "III", "2", "3"),
  tol = sqrt(.Machine$double.eps),
  ...
)
}
\arguments{
\item{mod}{(\code{mmrm})\cr the fitted MMRM.}

\item{type}{(\code{string}) type of test, "II", "III", '2', or '3'.}

\item{tol}{(\code{numeric}) threshold blow which values are treated as 0.}

\item{...}{not used.}
}
\description{
Conduct type II/III hypothesis testing on the MMRM fit results.
}
\details{
\code{Anova} will return \code{anova} object with one row per variable and columns
\verb{Num Df}(numerator degrees of freedom), \verb{Denom Df}(denominator degrees of freedom),
\verb{F Statistic} and \verb{Pr(>=F)}.
}
\keyword{internal}
