% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_package_version}
\alias{check_package_version}
\title{Check suggested dependency against version requirements}
\usage{
check_package_version(pkg, ver = c(NA, NA))
}
\arguments{
\item{pkg}{A package name}

\item{ver}{A vector of length two, specifying minimum and maximum (inclusive)
versions for interoperability.}
}
\value{
A logical (invisibly) indicating whether the loaded package meets
the version requirements. A warning is emitted otherwise.
}
\description{
Check suggested dependency against version requirements
}
\keyword{internal}
