% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_package_version}
\alias{check_package_version}
\title{Check Suggested Dependency Against Version Requirements}
\usage{
check_package_version(pkg, ver = c(NA_character_, NA_character_))
}
\arguments{
\item{pkg}{(\code{string})\cr package name.}

\item{ver}{(\code{character})\cr of length 2 whose elements can be provided to
\code{\link[=numeric_version]{numeric_version()}}, representing a minimum and maximum (inclusive) version
requirement for interoperability. When \code{NA}, no version requirement is
imposed. Defaults to no version requirement.}
}
\value{
A logical (invisibly) indicating whether the loaded package meets
the version requirements. A warning is emitted otherwise.
}
\description{
Check Suggested Dependency Against Version Requirements
}
\keyword{internal}
