% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{component}
\alias{component}
\title{Component Access for \code{mmrm_tmb} Objects}
\usage{
component(
  object,
  name = c("cov_type", "n_theta", "n_subjects", "n_timepoints", "n_obs", "vcov",
    "varcor", "formula", "dataset", "reml", "convergence", "evaluations", "conv_message",
    "call", "theta_est", "beta_est", "x_matrix", "y_vector", "neg_log_lik", "jac_list",
    "theta_vcov")
)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM.}

\item{name}{(\code{character})\cr the component(s) to be retrieved.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Available \code{component()} names are as follows:
\itemize{
\item \code{call}: low-level function call which generated the model.
\item \code{formula}: model formula.
\item \code{dataset}: data set name.
\item \code{cov_type}: covariance structure type.
\item \code{n_theta}: number of parameters.
\item \code{n_subjects}: number of subjects.
\item \code{n_timepoints}: number of modeled time points.
\item \code{n_obs}: total number of observations.
\item \code{reml}: was REML used (ML was used if \code{FALSE}).
\item \code{neg_log_lik}: negative log likelihood.
\item \code{convergence}: convergence code from optimizer.
\item \code{conv_message}: message accompanying the convergence code.
\item \code{evaluations}: number of function evaluations for optimization.
\item \code{vcov}: estimated variance-covariance matrix of coefficients.
\item \code{varcor}: estimated covariance matrix for residuals.
\item \code{theta_est}: estimated variance parameters.
\item \code{beta_est}: estimated coefficients.
\item \code{theta_vcov}:  estimated variance-covariance matrix of variance parameters.
\item \code{x_matrix}: design matrix used.
\item \code{y_vector}: response vector used.
\item \code{jac_list}: Jacobian, see  \code{\link[=h_jac_list]{h_jac_list()}} for details.
}
}
\seealso{
\code{\link[lme4]{getME.merMod}}

\code{\link[glmmTMB]{getME.glmmTMB}}
}
