% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\docType{data}
\name{COV_TYPES}
\alias{COV_TYPES}
\title{Covariance Type Database}
\format{
A data frame with 5 variables and one record per covariance type:

\describe{
\item{name}{
The long-form name of the covariance structure type
}
\item{abbr}{
The abbreviated name of the covariance structure type
}
\item{habbr}{
The abbreviated name of the heterogeneous version of a covariance
structure type (The abbreviated name (\code{abbr}) with a trailing \code{"h"} if
the structure has a heterogeneous implementation or \code{NA} otherwise).
}
\item{heterogeneous}{
A logical value indicating whether the covariance structure has a
heterogeneous counterpart.
}
\item{spatial}{
A logical value indicating whether the covariance structure is spatial.
}
}
}
\usage{
COV_TYPES
}
\description{
An internal constant for covariance type information.
}
\keyword{internal}
