% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_struct.R
\name{covariance_types}
\alias{covariance_types}
\alias{cov_types}
\title{Covariance Types}
\usage{
cov_types(
  form = c("name", "abbr", "habbr"),
  filter = c("heterogeneous", "spatial")
)
}
\arguments{
\item{form}{(\code{character})\cr Covariance structure type name form. One or
more of \code{"name"}, \code{"abbr"} (abbreviation), or \code{"habbr"} (heterogeneous
abbreviation).}

\item{filter}{(\code{character})\cr Covariance structure type filter. One or
more of \code{"heterogeneous"} or \code{"spatial"}.}
}
\value{
A character vector of accepted covariance structure type names and
abbreviations.
}
\description{
Covariance Types
}
\section{Abbreviations for Covariance Structures}{

\subsection{Common Covariance Structures:}{

\tabular{clll}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{   \\strong\{Structure\}
}\if{html}{\out{</div>}}

\tab \strong{Description}
\tab \strong{Parameters}
\tab \strong{\eqn{(i, j)} element}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   ad
}\if{html}{\out{</div>}}

\tab Ante-dependence
\tab \eqn{m}
\tab \eqn{\sigma^{2}\prod_{k=i}^{j-1}\rho_{k}}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   adh
}\if{html}{\out{</div>}}

\tab Heterogeneous ante-dependence
\tab \eqn{2m-1}
\tab \eqn{\sigma_{i}\sigma_{j}\prod_{k=i}^{j-1}\rho_{k}}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   ar1
}\if{html}{\out{</div>}}

\tab First-order auto-regressive
\tab \eqn{2}
\tab \eqn{\sigma^{2}\rho^{\left \vert {i-j} \right \vert}}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   ar1h
}\if{html}{\out{</div>}}

\tab Heterogeneous first-order auto-regressive
\tab \eqn{m+1}
\tab \eqn{\sigma_{i}\sigma_{j}\rho^{\left \vert {i-j} \right \vert}}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   cs
}\if{html}{\out{</div>}}

\tab Compound symmetry
\tab \eqn{2}
\tab \eqn{\sigma^{2}\left[ \rho I(i \neq j)+I(i=j) \right]}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   csh
}\if{html}{\out{</div>}}

\tab Heterogeneous compound symmetry
\tab \eqn{m+1}
\tab \eqn{\sigma_{i}\sigma_{j}\left[ \rho I(i \neq j)+I(i=j) \right]}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   toep
}\if{html}{\out{</div>}}

\tab Toeplitz
\tab \eqn{m}
\tab \eqn{\sigma_{\left \vert {i-j} \right \vert +1}}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   toeph
}\if{html}{\out{</div>}}

\tab Heterogeneous Toeplitz
\tab \eqn{2m-1}
\tab \eqn{\sigma_{i}\sigma_{j}\rho_{\left \vert {i-j} \right \vert}}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   us
}\if{html}{\out{</div>}}

\tab Unstructured
\tab \eqn{m(m+1)/2}
\tab \eqn{\sigma_{ij}}

}

where \eqn{i} and \eqn{j} denote \eqn{i}-th and \eqn{j}-th time points,
respectively, out of total \eqn{m} time points, \eqn{1 \leq i, j \leq m}.

Note the \strong{ante-dependence} covariance structure in this package refers to
homogeneous ante-dependence, while the ante-dependence covariance structure
from SAS \verb{PROC MIXED} refers to heterogeneous ante-dependence and the
homogeneous version is not available in SAS.
}

\subsection{Spatial Covariance structures:}{

\tabular{clll}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{   \\strong\{Structure\}
}\if{html}{\out{</div>}}

\tab \strong{Description}
\tab \strong{Parameters}
\tab \strong{\eqn{(i, j)} element}
\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{   sp_exp
}\if{html}{\out{</div>}}

\tab spatial exponential
\tab \eqn{2}
\tab \eqn{\sigma^{2}\rho^{-d_{ij}}}

}

where \eqn{d_{ij}} denotes the Euclidean distance between time points
\eqn{i} and \eqn{j}.
}
}

\seealso{
Other cov_struct: 
\code{\link{as.cov_struct}()},
\code{\link{cov_struct}()}
}
\concept{cov_struct}
