% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{emit_tidymodels_register_msg}
\alias{emit_tidymodels_register_msg}
\title{Format a Message to Emit When Tidymodels is Loaded}
\usage{
emit_tidymodels_register_msg()
}
\value{
A character message to emit. Either a ansi-formatted cli output if
package 'cli' is available or a plain-text message otherwise.
}
\description{
Format a Message to Emit When Tidymodels is Loaded
}
\keyword{internal}
