% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emp_start}
\alias{emp_start}
\title{Empirical Starting Value}
\usage{
emp_start(
  cov_type,
  full_frame,
  model_formula,
  group_var,
  visit_var,
  subject_var,
  ...
)
}
\description{
Empirical Starting Value
}
\details{
This \code{emp_start} only works for unstructured covariance structure.
}
