% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emp_start}
\alias{emp_start}
\title{Empirical Starting Value}
\usage{
emp_start(
  full_frame,
  model_formula,
  visit_var,
  subject_var,
  subject_groups,
  ...
)
}
\arguments{
\item{full_frame}{(\code{data.frame})\cr data used for model fitting.}

\item{model_formula}{(\code{formula})\cr the formula in mmrm model without covariance structure part.}

\item{visit_var}{(\code{string})\cr visit variable.}

\item{subject_var}{(\code{string})\cr subject id variable.}

\item{subject_groups}{(\code{factor})\cr subject group assignment.}

\item{...}{not used.}
}
\value{
A numeric vector of starting values.
}
\description{
Obtain empirical start value for unstructured covariance
}
\details{
This \code{emp_start} only works for unstructured covariance structure.
}
