% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{formula_rhs}
\alias{formula_rhs}
\title{Extract Right-Hand-Side (rhs) from Formula}
\usage{
formula_rhs(f)
}
\arguments{
\item{f}{(\code{formula})\cr a formula.}
}
\value{
A formula without a response, derived from the right-hand-side of the
formula, \code{f}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{formula_rhs(a ~ b + c)
formula_rhs(~ b + c)
}\if{html}{\out{</div>}}
}
\description{
Extract Right-Hand-Side (rhs) from Formula
}
\keyword{internal}
