% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_check_columns_nested}
\alias{h_check_columns_nested}
\alias{h_check_fits_all_data_same}
\title{Predicates Indicating Whether Datasets Contain the Same Observations}
\usage{
h_check_columns_nested(data_basic, data_augmented)

h_check_fits_all_data_same(fits)
}
\arguments{
\item{data_basic, data_augmented}{(\code{data.frame})\cr data frames to be
compared.}

\item{fits}{(\code{list})\cr list of \code{mmrm} fits.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Indicates whether adjacent datasets have the same rows and
nested columns (i.e., each dataset's columns are a subset of the next
dataset's columns).

For efficiency, the inspection takes place in this order:
\enumerate{
\item \code{FALSE} is returned early if not all datasets have the same number of
rows.
\item \code{FALSE} is returned early if a dataset has a column not in the next
dataset.
\item The columns in common among adjacent datasets are sorted and compared
using \code{\link[=all.equal]{all.equal()}} with \code{check.attributes = FALSE}.
}
}
\section{Functions}{
\itemize{
\item \code{h_check_columns_nested()}: checks two data sets to see if they
contain the same observations. The second data set must be the same as the
first data set, optionally with additional columns (ignoring attributes).

\item \code{h_check_fits_all_data_same()}: checks whether each fitted model's dataset
contains the same observations as the next model's dataset. Essentially,
\code{h_check_columns_nested()} is performed on the datasets of each adjacent
pair of \code{fits}.

}}
\keyword{internal}
