% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_covbeta_fun}
\alias{h_covbeta_fun}
\title{Covariance Matrix for Coefficients Given Variance Parameters}
\usage{
h_covbeta_fun(model)
}
\arguments{
\item{model}{(\code{mmrm_tmb})\cr initial model fit.}
}
\value{
Function with argument \code{theta} that calculates the covariance matrix
for the coefficient vector \code{beta}.
}
\description{
Covariance Matrix for Coefficients Given Variance Parameters
}
\keyword{internal}
