% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_covbeta_fun}
\alias{h_covbeta_fun}
\title{Covariance Matrix for Coefficients Given Variance Parameters}
\usage{
h_covbeta_fun(model)
}
\arguments{
\item{model}{(\code{mmrm_tmb})\cr initial model fit.}
}
\value{
Function with argument \code{theta} that calculates the covariance matrix
for the coefficient vector \code{beta}.
}
\description{
Obtain the covariance matrix function which accepts variance parameters as input.
Used in \code{\link[=h_jac_list]{h_jac_list()}} to to calculate the Jacobian matrix using numerical differentiation.
}
\keyword{internal}
