% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_default_value}
\alias{h_default_value}
\title{Default Value on NULL
Return default value when first argument is NULL.}
\usage{
h_default_value(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}
}
\description{
Default Value on NULL
Return default value when first argument is NULL.
}
\details{
If \code{x} is NULL, returns \code{y}. Otherwise return \code{x}.
}
\keyword{internal}
