% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_df_1d_kr}
\alias{h_df_1d_kr}
\title{Calculation of Kenward-Roger Degrees of Freedom for One-Dimensional Contrast}
\usage{
h_df_1d_kr(object, contrast, linear = FALSE)
}
\arguments{
\item{object}{(\code{mmrm})\cr object created by \code{\link[=mmrm]{mmrm()}} with Kenward-Roger(-Linear) method.}

\item{contrast}{(\code{numeric})\cr contrast vector.}

\item{linear}{(\code{logical})\cr whether to use linear Kenward-Roger approximation.}
}
\value{
List with \code{est}, \code{se}, \code{df}, \code{t_stat} and \code{p_val}.
}
\description{
Calculates the estimate, adjusted standard error, Kenward-Roger degree of freedom,
t statistic and p-value for one-dimensional contrast. Used in \code{\link[=df_1d]{df_1d()}} if method is
"Kenward-Roger" or "Kenward-Roger-Linear".
}
\keyword{internal}
