% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_df_md_from_1d}
\alias{h_df_md_from_1d}
\title{Creating F-Statistic Results from One-Dimensional Contrast}
\usage{
h_df_md_from_1d(object, contrast)
}
\arguments{
\item{object}{(\code{mmrm})\cr model fit.}

\item{contrast}{(\code{numeric})\cr one-dimensional contrast.}
}
\value{
The one-dimensional results are calculated and then returned as per
\code{\link[=h_df_md_list]{h_df_md_list()}}.
}
\description{
Creating F-Statistic Results from One-Dimensional Contrast
}
\keyword{internal}
