% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_df_md_kr}
\alias{h_df_md_kr}
\title{Calculation of Kenward-Roger Degrees of Freedom for Multi-Dimensional Contrast}
\usage{
h_df_md_kr(object, contrast, linear = FALSE)
}
\arguments{
\item{object}{\code{mmrm} object}

\item{contrast}{\code{matrix} contrast matrix}

\item{linear}{\code{logical}\cr whether to use linear Kenward-Roger approximation}
}
\value{
List with \code{num_df}, \code{denom_df}, \code{f_stat} and \code{p_val} (2-sided p-value).
}
\description{
Calculation of Kenward-Roger Degrees of Freedom for Multi-Dimensional Contrast
}
\keyword{internal}
