% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_df_md_kr}
\alias{h_df_md_kr}
\title{Calculation of Kenward-Roger Degrees of Freedom for Multi-Dimensional Contrast}
\usage{
h_df_md_kr(object, contrast, linear = FALSE)
}
\arguments{
\item{object}{(\code{mmrm})\cr object.}

\item{contrast}{(\code{matrix})\cr contrast matrix.}

\item{linear}{(\code{logical})\cr whether to use linear Kenward-Roger approximation.}
}
\value{
List with \code{num_df}, \code{denom_df}, \code{f_stat} and \code{p_val} (2-sided p-value).
}
\description{
Calculates the Kenward-Roger degree of freedom, F statistic and p value for multi-dimensional contrast.
Used in \code{\link[=df_md]{df_md()}} if method is "Kenward-Roger" or "Kenward-Roger-Linear".
}
\keyword{internal}
