% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_df_md_list}
\alias{h_df_md_list}
\title{Creating Results List for Multi-Dimensional Contrast}
\usage{
h_df_md_list(f_stat, num_df, denom_df)
}
\arguments{
\item{f_stat}{(\code{number})\cr F-statistic.}

\item{num_df}{(\code{number})\cr numerator degrees of freedom.}

\item{denom_df}{(\code{number})\cr denominator degrees of freedom.}
}
\value{
List with \code{num_df}, \code{denom_df}, \code{f_stat} and \code{p_val} (2-sided p-value).
}
\description{
Calculate the p-value using the F statistic and the numerator/denominator
degree of freedom and return a list.
}
\keyword{internal}
