% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{h_df_md_res}
\alias{h_df_md_res}
\title{Calculation of Residual Degrees of Freedom for Multi-Dimensional Contrast}
\usage{
h_df_md_res(object, contrast)
}
\arguments{
\item{object}{(\code{mmrm})\cr the MMRM fit.}

\item{contrast}{(\code{matrix})\cr numeric contrast matrix, if given a \code{numeric}
then this is coerced to a row vector. Note that this should not include
elements for singular coefficient estimates, i.e. only refer to the
actually estimated coefficients.}
}
\value{
List with \code{num_df}, \code{denom_df}, \code{f_stat} and \code{p_val} (2-sided p-value).
}
\description{
Used in \code{\link[=df_md]{df_md()}} if method is "Residual".
}
\keyword{internal}
