% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_factor_ref}
\alias{h_factor_ref}
\title{Convert Character to Factor Following Reference}
\usage{
h_factor_ref(x, ref, var_name = vname(x))
}
\arguments{
\item{x}{(\code{character} or \code{factor})\cr  input.}

\item{ref}{(\code{factor})\cr reference.}

\item{var_name}{(\code{string})\cr variable name of input \code{x}.}
}
\value{
Factor vector with updated levels.
}
\description{
Convert Character to Factor Following Reference
}
\details{
Use \code{ref} to convert \code{x} into factor with the same levels.
This is needed even if \code{x} and \code{ref} are both \code{character} because
in \code{model.matrix} if \code{x} only has one level there could be errors.
}
\keyword{internal}
