% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_factor_ref_data}
\alias{h_factor_ref_data}
\title{Convert Character to Factor Following Reference \code{MMRM} Fit.}
\usage{
h_factor_ref_data(object, data)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM object.}

\item{data}{(\code{data.frame})\cr input data.}
}
\value{
Data frame with updated levels in specified columns.
}
\description{
Convert Character to Factor Following Reference \code{MMRM} Fit.
}
\details{
Use fitted mmrm object to convert input data frame whose factors
are of the same levels as the reference fitted object.
}
\keyword{internal}
