% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_get_cov_default}
\alias{h_get_cov_default}
\title{Obtain Default Covariance Method}
\usage{
h_get_cov_default(
  method = c("Satterthwaite", "Kenward-Roger", "Residual", "Between-Within")
)
}
\arguments{
\item{method}{(\code{string})\cr degrees of freedom method.}
}
\value{
String of the default covariance method.
}
\description{
Obtain the default covariance method depending on
the degrees of freedom method used.
}
\details{
The default covariance method is different for different degrees of freedom method.
For "Satterthwaite" or "Between-Within", "Asymptotic" is returned.
For "Kenward-Roger" only, "Kenward-Roger" is returned.
For "Residual" only, "Empirical" is returned.
}
\keyword{internal}
