% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical.R
\name{h_get_empirical}
\alias{h_get_empirical}
\title{Obtain Empirical/Jackknife Covariance.}
\usage{
h_get_empirical(tmb_data, theta, beta, beta_vcov, jackknife)
}
\arguments{
\item{tmb_data}{(\code{mmrm_tmb_data})\cr produced by \code{\link[=h_mmrm_tmb_data]{h_mmrm_tmb_data()}}.}

\item{theta}{(\code{numeric})\cr theta estimate.}

\item{beta}{(\code{numeric})\cr beta estimate.}

\item{beta_vcov}{(\code{matrix})\cr covariance of beta estimate.}

\item{jackknife}{(\code{flag})\cr indicator of whether Jackknife covariance is returned.}
}
\value{
The matrix of covariance matrix of beta.
}
\description{
Obtain the empirical or Jackknife covariance for \eqn{\beta}.
Used in \code{mmrm} fitting if method is "Empirical" or "Empirical-Jackknife".
}
\keyword{internal}
