% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kr_comp.R
\name{h_get_kr_comp}
\alias{h_get_kr_comp}
\title{obtain kr component}
\usage{
h_get_kr_comp(tmb_data, theta)
}
\arguments{
\item{tmb_data}{(\code{mmrm_tmb_data})\cr produced by \code{\link[=h_mmrm_tmb_data]{h_mmrm_tmb_data()}}.}

\item{theta}{numeric vector of the theta estimate.}
}
\description{
obtain kr component
}
\details{
the function retuns a named list, $P$, $Q$ and $R$, which corresponds to the
paper in 1997. The matrices are stacked in columns so that $P$, $Q$ and $R$ has the same
column number(number of beta parameters). The number of rows, is dependent on
the total number of theta and number of groups, if the fit is a grouped mmrm.
For $P$ matrix, it is stacked sequentially. For $Q$ and $R$ matrix, it is stacked so
that the $Q_{ij}$ and $R_{ij}$ is stacked from $j$ then to $i$, i.e. $R_{i1}$, $R_{i2}$, etc.
$Q$ and $R$ only contains within group results and intra-group results should be all zero matrices
so they are not stacked in the result.
}
\keyword{internal}
