% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_get_prediction}
\alias{h_get_prediction}
\title{Get Prediction}
\usage{
h_get_prediction(tmb_data, theta, beta, beta_vcov)
}
\arguments{
\item{tmb_data}{(\code{mmrm_tmb_data})\cr object.}

\item{theta}{(\code{numeric})\cr theta value.}

\item{beta}{(\code{numeric})\cr beta value.}

\item{beta_vcov}{(\code{matrix})\cr beta_vcov matrix.}
}
\value{
List with:
\itemize{
\item \code{prediction}: Matrix with columns \code{fit}, \code{conf_var}, and \code{var}.
\item \code{covariance}: List with subject specific covariance matrices.
\item \code{index}: List of zero-based subject indices.
}
}
\description{
Get predictions with given \code{data}, \code{theta}, \code{beta}, \code{beta_vcov}.
}
\details{
See \code{predict} function in \code{predict.cpp} which is called internally.
}
\keyword{internal}
