% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_get_prediction_variance}
\alias{h_get_prediction_variance}
\title{Get Prediction Variance}
\usage{
h_get_prediction_variance(object, nsim, tmb_data)
}
\arguments{
\item{object}{(\code{mmrm_tmb})\cr the fitted MMRM.}

\item{nsim}{(\code{count})\cr number of samples.}

\item{tmb_data}{(\code{mmrm_tmb_data})\cr object.}
}
\description{
Get prediction variance with given fit, \code{tmb_data} with the Monte Carlo sampling method.
}
\keyword{internal}
