% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{h_get_sim_per_subj}
\alias{h_get_sim_per_subj}
\title{Get simulated values by patient.}
\usage{
h_get_sim_per_subj(predict_res, nsub, nsim)
}
\arguments{
\item{predict_res}{(\code{list})\cr from \code{\link[=h_get_prediction]{h_get_prediction()}}.}

\item{nsub}{(\code{count})\cr number of subjects.}

\item{nsim}{(\code{count})\cr number of values to simulate.}
}
\description{
Get simulated values by patient.
}
\keyword{internal}
