% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_get_theta_from_cov}
\alias{h_get_theta_from_cov}
\title{Obtain Theta from Covariance Matrix}
\usage{
h_get_theta_from_cov(covariance)
}
\arguments{
\item{covariance}{(\code{matrix}) of covariance matrix values.}
}
\description{
Obtain unstructured theta from covariance matrix.
}
\details{
If the covariance matrix has \code{NA} in some of the elements, they will be replaced by
0 (non-diagonal) and 1 (diagonal). This ensures that the matrix is positive definite.
}
\keyword{internal}
