% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_jac_list}
\alias{h_jac_list}
\title{Obtain List of Jacobian Matrix Entries for Covariance Matrix}
\usage{
h_jac_list(tmb_data, theta_est, beta_vcov)
}
\arguments{
\item{tmb_data}{(\code{mmrm_tmb_data})\cr produced by \code{\link[=h_mmrm_tmb_data]{h_mmrm_tmb_data()}}.}

\item{theta_est}{(\code{numeric})\cr variance parameters point estimate.}

\item{beta_vcov}{(\code{matrix})\cr vairance covariance matrix of coefficients.}
}
\value{
List with one element per variance parameter containing a matrix
of the same dimensions as the covariance matrix. The values are the derivatives
with regards to this variance parameter.
}
\description{
Obtain the Jacobian matrices given the covariance function and variance parameters.
}
\keyword{internal}
