% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{h_jac_list}
\alias{h_jac_list}
\title{Obtain List of Jacobian Matrix Entries for Covariance Matrix}
\usage{
h_jac_list(covbeta_fun, theta_est)
}
\arguments{
\item{covbeta_fun}{(\code{function})\cr function calculating the covariance
matrix of coefficients given variance parameters (\code{theta}), see
\code{\link[=h_covbeta_fun]{h_covbeta_fun()}} to obtain this from a \code{mmrm_tmb} object.}

\item{theta_est}{(\code{numeric})\cr variance parameters point estimate.}
}
\value{
List with one element per variance parameter containing a matrix
of the same dimensions as the covariance matrix. The values are the derivatives
with regards to this variance parameter.
}
\description{
Obtain the Jacobian matrices given the covariance function and variance parameters.
}
\keyword{internal}
