% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_kr_df}
\alias{h_kr_df}
\title{obtain the adjusted Kenward-Roger degree of freedom}
\usage{
h_kr_df(v0, l, w, p)
}
\arguments{
\item{v0}{unadjusted covariance matrix}

\item{l}{linear combination matrix}

\item{w}{hessian matrix}

\item{p}{P matrix from \code{h_get_kr_comp}}
}
\value{
List with \code{m} and \code{lambda}
}
\description{
obtain the adjusted Kenward-Roger degree of freedom
}
\keyword{internal}
