% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_kr_df}
\alias{h_kr_df}
\title{Obtain the Adjusted Kenward-Roger Degree of Freedom}
\usage{
h_kr_df(v0, l, w, p)
}
\arguments{
\item{v0}{(\code{matrix})\cr unadjusted covariance matrix}

\item{l}{(\code{matrix})\cr linear combination matrix}

\item{w}{(\code{matrix})\cr hessian matrix}

\item{p}{(\code{matrix})\cr P matrix from \code{h_get_kr_comp}}
}
\value{
Named list with elements:
\itemize{
\item \code{m}: \code{numeric} degree of freedom.
\item \code{lambda}: \code{numeric} f statistic scaler.
}
}
\description{
Obtain the Adjusted Kenward-Roger Degree of Freedom
}
\keyword{internal}
