% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_kr_df}
\alias{h_kr_df}
\title{Obtain the Adjusted Kenward-Roger degrees of freedom}
\usage{
h_kr_df(v0, l, w, p)
}
\arguments{
\item{v0}{(\code{matrix})\cr unadjusted covariance matrix.}

\item{l}{(\code{matrix})\cr linear combination matrix.}

\item{w}{(\code{matrix})\cr hessian matrix.}

\item{p}{(\code{matrix})\cr P matrix from \code{\link[=h_get_kr_comp]{h_get_kr_comp()}}.}
}
\value{
Named list with elements:
\itemize{
\item \code{m}: \code{numeric} degrees of freedom.
\item \code{lambda}: \code{numeric} F statistic scale parameter.
}
}
\description{
Obtains the adjusted Kenward-Roger degrees of freedom and F statistic scale parameter.
Used in \code{\link[=h_df_md_kr]{h_df_md_kr()}} or \link{h_df_1d_kr}.
}
\keyword{internal}
