% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_assert_start}
\alias{h_mmrm_tmb_assert_start}
\title{Asserting Sane Start Values for \code{TMB} Fit}
\usage{
h_mmrm_tmb_assert_start(tmb_object)
}
\arguments{
\item{tmb_object}{(\code{list})\cr created with \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}
}
\value{
Nothing, only used for assertions.
}
\description{
Asserting Sane Start Values for \code{TMB} Fit
}
\keyword{internal}
