% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_data}
\alias{h_mmrm_tmb_data}
\title{Data for \code{TMB} Fit}
\usage{
h_mmrm_tmb_data(formula_parts, data, reml)
}
\arguments{
\item{formula_parts}{(\code{mmrm_tmb_formula_parts})\cr list with formula parts
from \code{\link[=h_mmrm_tmb_formula_parts]{h_mmrm_tmb_formula_parts()}}.}

\item{data}{(\code{data.frame})\cr which contains variables used in \code{formula_parts}.}

\item{reml}{(\code{flag})\cr whether restricted maximum likelihood (REML) estimation is used,
otherwise maximum likelihood (ML) is used.}
}
\value{
List of class \code{mmrm_tmb_data} with elements:
\itemize{
\item \code{x_matrix}: \code{matrix} with \code{n} rows and \code{p} columns specifying the overall design matrix.
\item \code{y_vector}: length \code{n} \code{numeric} specifying the overall response vector.
\item \code{visits_zero_inds}: length \code{n} \code{integer} containing zero-based visits indices.
\item \code{n_visits}: \code{int} with the number of visits, which is the dimension of the
covariance matrix.
\item \code{n_subjects}: \code{int} with the number of subjects.
\item \code{subject_zero_inds}: length \code{n_subjects} \code{integer} containing the zero-based start
indices for each subject.
\item \code{subject_n_visits}: length \code{n_subjects} \code{integer} containing the number of
observed visits for each subjects. So the sum of this vector equals \code{n}.
\item \code{corr_type}: \code{int} specifying the correlation type.
\item \code{reml}: \code{int} specifying whether REML estimation is used (1), otherwise ML (0).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Note that the \code{subject_var} must not be factor but can also be character.
If it is character, then it will be converted to factor internally. Here
the levels will be the unique values, sorted alphabetically and numerically if there
is a common string prefix of numbers in the character elements. For full control
on the order please use a factor.
}
\examples{
formula <- FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
formula_parts <- h_mmrm_tmb_formula_parts(formula)
tmb_data <- h_mmrm_tmb_data(formula_parts, fev_data, reml = FALSE)
}
