% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{h_mmrm_tmb_parameters}
\alias{h_mmrm_tmb_parameters}
\title{Start Parameters for \code{TMB} Fit}
\usage{
h_mmrm_tmb_parameters(formula_parts, tmb_data, start, n_groups = 1L)
}
\arguments{
\item{formula_parts}{(\code{mmrm_tmb_formula_parts})\cr produced by
\code{\link[=h_mmrm_tmb_formula_parts]{h_mmrm_tmb_formula_parts()}}.}

\item{tmb_data}{(\code{mmrm_tmb_data})\cr produced by \code{\link[=h_mmrm_tmb_data]{h_mmrm_tmb_data()}}.}

\item{start}{(\code{numeric} or \code{NULL})\cr optional start values for variance
parameters.}

\item{n_groups}{(\code{int})\cr number of groups.}
}
\value{
List with element \code{theta} containing the start values for the variance
parameters.
}
\description{
Start Parameters for \code{TMB} Fit
}
\keyword{internal}
