% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-methods.R
\name{h_print_cov}
\alias{h_print_cov}
\title{Printing MMRM Covariance Type}
\usage{
h_print_cov(cov_type, n_theta, n_groups)
}
\arguments{
\item{cov_type}{(\code{string})\cr covariance structure abbreviation.}

\item{n_theta}{(\code{count})\cr number of variance parameters.}

\item{n_groups}{(\code{count})\cr number of groups.}
}
\description{
This is used in \code{\link[=print.summary.mmrm]{print.summary.mmrm()}}.
}
\keyword{internal}
