% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{h_summarize_all_fits}
\alias{h_summarize_all_fits}
\title{Summarizing List of Fits}
\usage{
h_summarize_all_fits(all_fits)
}
\arguments{
\item{all_fits}{(\code{list} of \code{mmrm_fit} or \code{try-error})\cr list of fits.}
}
\value{
List with \code{warnings}, \code{messages}, \code{log_liks} and \code{converged} results.
}
\description{
Summarizing List of Fits
}
\keyword{internal}
