% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_var_adj}
\alias{h_var_adj}
\title{obtain the adjusted covariance matrix}
\usage{
h_var_adj(v, w, p, q, r, order = 2L)
}
\arguments{
\item{v}{unadjusted covariance matrix}

\item{w}{hessian matrix}

\item{p}{P matrix from \code{h_get_kr_comp}}

\item{q}{Q matrix from \code{h_get_kr_comp}}

\item{r}{R matrix from \code{h_get_kr_comp}}

\item{order}{the order of the Kenward-Roger approximation. Only 1L or 2L supported.}
}
\description{
obtain the adjusted covariance matrix
}
\keyword{internal}
