% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenwardroger.R
\name{h_var_adj}
\alias{h_var_adj}
\title{obtain the adjusted covariance matrix}
\usage{
h_var_adj(v, w, p, q, r, linear = FALSE)
}
\arguments{
\item{v}{unadjusted covariance matrix}

\item{w}{hessian matrix}

\item{p}{P matrix from \code{h_get_kr_comp}}

\item{q}{Q matrix from \code{h_get_kr_comp}}

\item{r}{R matrix from \code{h_get_kr_comp}}

\item{linear}{\code{logical}\cr whether to use linear Kenward-Roger approximation}
}
\value{
matrix of adjusted covariance matrix
}
\description{
obtain the adjusted covariance matrix
}
\keyword{internal}
