% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mmrm-package}
\alias{mmrm}
\alias{mmrm-package}
\title{mmrm: Mixed model repeated measures and its extension}
\description{
Mixed model repeated measures and its extension.
}
\author{
\strong{Maintainer}: Siyi Liu \email{sliu48@ncsu.edu}

Authors:
\itemize{
  \item Yilong Zhang \email{zhanyilo@merck.com}
}

}
\keyword{internal}
