% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-package.R
\docType{package}
\name{mmrm-package}
\alias{mmrm-package}
\alias{_PACKAGE}
\title{\code{mmrm} Package}
\description{
\code{mmrm} implements mixed models with repeated measures (MMRM) in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/mmrm/}
  \item Report bugs at \url{https://github.com/openpharma/mmrm/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}

Authors:
\itemize{
  \item Julia Dedic \email{julia.dedic@roche.com}
  \item Brian Matthew Lang \email{brian.lang@msd.com}
  \item Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
  \item Craig Gower-Page \email{craig.gower-page@roche.com}
  \item Alessandro Noci \email{alessandro.noci@roche.com}
}

}
