% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mmrm-package}
\alias{mmrm}
\alias{mmrm-package}
\title{\code{mmrm} Package}
\description{
\code{mmrm} implements mixed models with repeated measures (MMRM) in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/mmrm/}
  \item Report bugs at \url{https://github.com/openpharma/mmrm/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}

}
