% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm-package.R
\docType{package}
\name{mmrm-package}
\alias{mmrm-package}
\alias{_PACKAGE}
\title{\code{mmrm} Package}
\description{
\code{mmrm} implements mixed models with repeated measures (MMRM) in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openpharma.github.io/mmrm/}
  \item Report bugs at \url{https://github.com/openpharma/mmrm/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com}

Authors:
\itemize{
  \item Julia Dedic \email{julia.dedic@roche.com}
  \item Doug Kelkhoff \email{doug.kelkhoff@roche.com}
  \item Kevin Kunzmann \email{kevin.kunzmann@boehringer-ingelheim.com}
  \item Brian Matthew Lang \email{brian.lang@msd.com}
  \item Liming Li \email{liming.li@roche.com}
  \item Christian Stock \email{christian.stock@boehringer-ingelheim.com}
  \item Ya Wang \email{ya.wang10@gilead.com}
  \item Dan James \email{dan.james@astrazeneca.com}
  \item Jonathan Sidi \email{yoni.sidi@sagerx.com}
  \item Daniel D. Sjoberg \email{sjobergd@gene.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
}

Other contributors:
\itemize{
  \item Craig Gower-Page \email{craig.gower-page@roche.com} [contributor]
  \item Boehringer Ingelheim Ltd. [copyright holder, funder]
  \item Gilead Sciences, Inc. [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Merck Sharp & Dohme, Inc. [copyright holder, funder]
  \item AstraZeneca plc [copyright holder, funder]
}

}
