% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm.R
\name{mmrm}
\alias{mmrm}
\title{Fits MMRM model using nlme::gls}
\usage{
mmrm(
  formula,
  time,
  subjects,
  data,
  cov_struct = c("unstructured", "autoregressive", "compound-symmetry"),
  control = nlme::glsControl(),
  stop_on_convergence = TRUE,
  return_all = FALSE,
  ...
)
}
\arguments{
\item{formula}{formula for the fixed effects structure of the model}

\item{time}{the time variable of the model}

\item{subjects}{the variable indicating unique subjects}

\item{data}{the data structure}

\item{cov_struct}{a sequence of covariance matrix specifications;
will fit a model starting with the first covariance
and check for convergence. If there are convergence
issues, it will iterate through the rest of the list
until the model converges.}

\item{control}{a glsControl object, see \link[nlme]{glsControl}}

\item{stop_on_convergence}{if TRUE, stop fitting models and
if FALSE, fits models with all specified
covariance structures regardless of
convergence of earlier model}

\item{return_all}{logical; if TRUE, will return all models that were fit,
if FALSE, only returns the first model that converged
or last model attempted (if all failed).}

\item{...}{additional arguments passed to \link[nlme]{gls}
(e.g., na.action)}
}
\value{
glsObject or list of glsObjects
           if return_all = FALSE, returns only the glsObject
             of the first model to converge or the last attempted model.
           if return_all = TRUE, returns list of all attempted models
}
\description{
Fits a Mixed Model Repeated Measures model (see Details).
The fixed effects structure is flexible -- user defined using a formula.
The random effects structure is hard-coded:
independent variance at each time point,
but residuals of individual subjects are correlated across timepoints.
}
\details{
TODO
}
\examples{
\donttest{
mmrm(outcome ~ baseline + group + time + baseline:time + group:time,
     time = "time",
     subjects = "subjectid",
     data = data)
}

}
