% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmrm_control}
\alias{mmrm_control}
\title{Control Parameters for Fitting an MMRM}
\usage{
mmrm_control(
  n_cores = 1L,
  method = c("Satterthwaite", "Kenward-Roger", "Kenward-Roger-Linear"),
  start = NULL,
  accept_singular = TRUE,
  ...,
  optimizers = h_get_optimizers(...)
)
}
\arguments{
\item{n_cores}{(\code{int})\cr number of cores to be used.}

\item{method}{(\code{character})\cr vector of covariance/degree of freedom method.}

\item{start}{(\code{numeric} or \code{NULL})\cr optional start values for variance
parameters.}

\item{accept_singular}{(\code{flag})\cr whether singular design matrices are reduced
to full rank automatically and additional coefficient estimates will be missing.}

\item{...}{Additional arguments passed to \code{\link[=h_get_optimizers]{h_get_optimizers()}}.}

\item{optimizers}{(\code{list})\cr of optimizers created from \code{\link[=h_get_optimizers]{h_get_optimizers()}}.}
}
\value{
List of class \code{mmrm_control} with the control parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
mmrm_control(
  optimizer_fun = stats::optim,
  optimizer_args = list(method = "L-BFGS-B")
)
}
