% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_eff_size.R
\name{mmrm_eff_size}
\alias{mmrm_eff_size}
\title{Calculate effect size for contrasts using estimated marginal means}
\usage{
mmrm_eff_size(mmrm_model, mmrm_emm, residual_error = TRUE, ...)
}
\arguments{
\item{mmrm_model}{the mmrm model(s) of type \code{mmrm}, \code{mmrmList}, or \code{mmrmCV}
used to calculate effect size}

\item{mmrm_emm}{estimated marginal means from and MMRM.
output of \link{mmrm_emmeans}}

\item{...}{additional arguments passed to \link[emmeans:eff_size]{emmeans::eff_size}}

\item{edf}{equivalent degrees of freedom for the residual variance.
See \link[emmeans:eff_size]{emmeans::eff_size} for details, by default NULL}
}
\value{
data.frame with effect sizes calculated by emmeans
}
\description{
Calculate timepoint by timepoint effect size for MMRM models as the
least-squares means / sqrt(residual variance) at each timepoint.
Calculated by running \link[emmeans:eff_size]{emmeans::eff_size} for each timepoint contrast,
where the timepoint by timepoint residuals come from the diagonals of \link{varcov}.
Uncertainty is estimated by \link[emmeans:eff_size]{emmeans::eff_size}.
}
