% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrm_emmeans.R
\name{mmrm_emmeans}
\alias{mmrm_emmeans}
\title{Estimated Marignal Means of MMRM models}
\usage{
mmrm_emmeans(...)
}
\arguments{
\item{...}{see \link[emmeans:emmeans]{emmeans::emmeans}}
}
\value{
see \link[emmeans:emmeans]{emmeans::emmeans}
}
\description{
Alias for \link[emmeans:emmeans]{emmeans::emmeans} that provides support
for using Kenward-Rogers degrees of freedom calculation.
Please see \link[emmeans:emmeans]{emmeans::emmeans} for further documentation.
}
\details{
Support for Kenward-Rogers degrees of freedom calculation
provided with i) a custom emm_basis function that overwrites the
emm_basis.glm provided by the emmeans package and
ii) addition of support for glsObject in the pbkrtest package
(currently in a development version
\href{https://github.com/gkane26/pbkrtest/tree/nlme}{gkane26/pbkrtest@nlme}).

\if{html}{\out{<div class="sourceCode">}}\preformatted{      Kenward-Rogers d.f. calculation is memory and computation time intensive.
      By default, running this method on a large dataset will throw an error to warn users.
      To bypass this error, please set the pbkrtest.limit argument to a number larger than
      the number of observations in the model.
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
mmrm_emmeans(object = mmrm_object,
             specs = pairwise ~ time | treatment,
             mode = "kenward",
             pbkrtest.limit = emmeans::get_emm_option("pbkrtest.limit"))
}

}
