% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{mmrm_tmb}
\alias{mmrm_tmb}
\title{Fitting an MMRM with \code{TMB}}
\usage{
mmrm_tmb(formula, data, start = NULL, control = h_mmrm_tmb_control())
}
\arguments{
\item{formula}{(\code{formula})\cr model formula with exactly one special term
specifying the visits within subjects, see details.}

\item{data}{(\code{data.frame})\cr}

\item{start}{(\code{list} or \code{NULL})\cr}

\item{control}{(\code{mmrm_tmb_control})\cr}
}
\value{

}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
The \code{formula} typically looks like:
\code{AVAL ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)}
so specifies response and covariates as usual, and exactly one special term
defines which correlation structure is used and what are the visit and
subject variables.
}
