% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-parsnip.R
\name{parsnip_add_mmrm}
\alias{parsnip_add_mmrm}
\title{Register \code{mmrm} For Use With \code{tidymodels}}
\usage{
parsnip_add_mmrm(quietly = FALSE)
}
\arguments{
\item{quietly}{logical: should progress and error messages be suppressed?}
}
\value{
A logical value indicating whether registration was successful.
}
\description{
Register \code{mmrm} For Use With \code{tidymodels}
}
\details{
We can use \code{parsnip::show_model_info("linear_reg")} to check the
registration with \code{parsnip} and thus the wider \code{tidymodels} ecosystem.
}
\keyword{internal}
