% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{register_on_load}
\alias{register_on_load}
\title{Helper Function for Registering Functionality With Suggests Packages}
\usage{
register_on_load(pkg, ver = c(NA, NA), callback, message = NULL)
}
\arguments{
\item{pkg}{(\code{character[1]})\cr
A package name.}

\item{ver}{(any object of length 2 whose elements can be provided to
\code{numeric_version()})\cr
A minimum and maximum (inclusive) version requirement for interoperability.
When \code{NA}, no version requirement is imposed. Defaults to no version
requirement.}

\item{callback}{(\code{function(...) ANY})\cr
A callback to execute upon package load. Note that no arguments are
passed to this function. Any necessary data must be provided upon
construction.}

\item{message}{(\code{NULL} or \code{character[1]})\cr
"An optional message to print after the callback is executed upon
successful registration.")}
}
\value{
A logical (invisibly) indicating whether registration was successful.
If not, a onLoad hook was set for the next time the package is loaded.
}
\description{
Helper Function for Registering Functionality With Suggests Packages
}
\keyword{internal}
