% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcov.R
\name{varcov}
\alias{varcov}
\title{Get Variance-Covariance matrix of MMRM model}
\usage{
varcov(model)
}
\arguments{
\item{model}{mmrmObject (output of \link{mmrm})}
}
\value{
matrix, the variance-covoariance matrix
}
\description{
Compute the variance-covariance matrix of an mmrmObject.
Copy of \href{https://cjangelo.github.io/SPR/reference/VarCov_gls.html}{VarCov_gls from the SPR package}
}
\seealso{
\link[SPR:VarCov_gls]{SPR::VarCov_gls}
}
